% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{assignFlyway}
\alias{assignFlyway}
\title{Assign flyway as new field}
\usage{
assignFlyway(x, state_field_name, flyway_field_name)
}
\arguments{
\item{x}{Tibble containing a field with two-letter state abbreviation}

\item{state_field_name}{Name of field with two-letter state abbreviations}

\item{flyway_field_name}{Name to use for flyway column assignment}
}
\description{
Internal helper function that assigns a new field for flyway name (e.g.,
"Atlantic Flyway") to a tibble containing an existing field for state
abbreviations (e.g., "MA").
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{helper functions}
