% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{badBirthDatesMessage}
\alias{badBirthDatesMessage}
\title{Bad birth dates message}
\usage{
badBirthDatesMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{badBirthDatesMessage} function returns a message for
\code{birth_date} values that do not have a range greater than 365 days.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
