% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{badRegYearMessage}
\alias{badRegYearMessage}
\title{Bad registration_yr message}
\usage{
badRegYearMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{badRegYearMessage} function returns a message for
\code{registration_yr} values that are not equal to REF_CURRENT_SEASON or
REF_CURRENT_SEASON + 1.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
