% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{badTitleMessage}
\alias{badTitleMessage}
\title{Bad title message}
\usage{
badTitleMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{badTitleMessage} function returns a message for
\code{title} values that are not assigned to the correct registrations.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
