% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile.R
\name{compile}
\alias{compile}
\title{Read in data}
\usage{
compile(path)
}
\arguments{
\item{path}{The path to the data .txt files}
}
\description{
Compile data from all state-exported text files by providing a path to the download directory. If the error "Directory contains file(s) with non-UTF-8 encoding." is shown, use \code{\link{compile_to_utf8}} instead.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
