% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_to_utf8.R
\name{compile_to_utf8}
\alias{compile_to_utf8}
\title{Read in data with non-utf-8 encoding}
\usage{
compile_to_utf8(path)
}
\arguments{
\item{path}{The path to the data .txt files}
}
\description{
Compile data from all state-exported text files by providing a path to the download directory
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
