% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correctEmail}
\alias{correctEmail}
\title{Correct email}
\usage{
correctEmail(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}}}
}
\description{
The internal \code{correctEmail} function is used inside of
\code{\link{correct}} to ...
}
\seealso{
Other correcting functions:
\code{\link[=correct]{correct()}},
\code{\link[=correctMiddleInitial]{correctMiddleInitial()}},
\code{\link[=correctSuffix]{correctSuffix()}},
\code{\link[=correctTitle]{correctTitle()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{correcting functions}
