% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correctMiddleInitial}
\alias{correctMiddleInitial}
\title{Correct middle initials}
\usage{
correctMiddleInitial(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}}}
}
\description{
The internal \code{correctMiddleInitial} function changes non-alphabetic
characters in the middle initial column to NA.
}
\seealso{
Other correcting functions:
\code{\link[=correct]{correct()}},
\code{\link[=correctEmail]{correctEmail()}},
\code{\link[=correctSuffix]{correctSuffix()}},
\code{\link[=correctTitle]{correctTitle()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{correcting functions}
