% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DF_TEST_MINI}
\alias{DF_TEST_MINI}
\title{Miniature fake HIP data}
\format{
A data frame with 1606 rows and 30 variables:
\describe{
\item{title}{Title code to indicate male or female hunter}
\item{firstname}{First name}
\item{middle}{Middle initial}
\item{lastname}{Last name}
\item{suffix}{Suffix}
\item{address}{Street or PO Box address}
\item{city}{City of residence}
\item{state}{State of residence}
\item{zip}{Zip code}
\item{birth_date}{Birth date}
\item{issue_date}{HIP registration date}
\item{hunt_mig_birds}{Code to indicate if hunter hunted migratory birds last season}
\item{ducks_bag}{Code to indicate duck harvest from last season}
\item{geese_bag}{Code to indicate goose harvest from last season}
\item{dove_bag}{Code to indicate dove harvest from last season}
\item{woodcock_bag}{Code to indicate woodcock harvest from last season}
\item{coots_snipe}{Code to indicate coot and snipe harvest from last season}
\item{rails_gallinules}{Code to indicate rail and gallinule harvest from last season}
\item{cranes}{Code to indicate crane harvest from last season}
\item{band_tailed_pigeon}{Code to indicate band-tailed pigeon harvest from last season}
\item{brant}{Code to indicate brant harvest from last season}
\item{seaducks}{Code to indicate sea duck harvest from last season}
\item{registration_yr}{Beginning year of the hunting season that HIP registration is valid for}
\item{email}{Email address}
\item{dl_state}{HIP registration state}
\item{dl_date}{Date of HIP file}
\item{source_file}{File name that the record is from}
\item{dl_cycle}{Download cycle}
\item{dl_key}{Key for dl_state and dl_date combination}
\item{record_key}{Key for each record}
}
}
\usage{
DF_TEST_MINI
}
\description{
Miniature fake HIP data generated by data-raw/create_fake_HIP_data.R
}
\keyword{datasets}
