% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{dlDateNAMessage}
\alias{dlDateNAMessage}
\title{Return message if there is an NA in dl_date}
\usage{
dlDateNAMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{dlDateNAMessage} function is used inside of \code{\link{readMessages}}
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
