% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateFinder}
\alias{duplicateFinder}
\title{Find duplicates}
\usage{
duplicateFinder(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Determine how many duplicate records are in the data and return a table.
}
\seealso{
Other deduplication functions:
\code{\link[=duplicateAllOnes]{duplicateAllOnes()}},
\code{\link[=duplicateAllOnesGroupSize]{duplicateAllOnesGroupSize()}},
\code{\link[=duplicateDecide]{duplicateDecide()}},
\code{\link[=duplicateFields]{duplicateFields()}},
\code{\link[=duplicateFix]{duplicateFix()}},
\code{\link[=duplicateID]{duplicateID()}},
\code{\link[=duplicateNewest]{duplicateNewest()}},
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=duplicateRecordType]{duplicateRecordType()}},
\code{\link[=duplicateSample]{duplicateSample()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
