% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateFix}
\alias{duplicateFix}
\title{Fix duplicates}
\usage{
duplicateFix(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Resolve duplicate HIP records.
}
\seealso{
Other deduplication functions:
\code{\link[=duplicateAllOnes]{duplicateAllOnes()}},
\code{\link[=duplicateAllOnesGroupSize]{duplicateAllOnesGroupSize()}},
\code{\link[=duplicateDecide]{duplicateDecide()}},
\code{\link[=duplicateFields]{duplicateFields()}},
\code{\link[=duplicateFinder]{duplicateFinder()}},
\code{\link[=duplicateID]{duplicateID()}},
\code{\link[=duplicateNewest]{duplicateNewest()}},
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=duplicateRecordType]{duplicateRecordType()}},
\code{\link[=duplicateSample]{duplicateSample()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
