% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateID}
\alias{duplicateID}
\title{Create a tibble of duplicates with an ID column}
\usage{
duplicateID(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Return a tibble of duplicates with a duplicate ID column identifying each
group of records.
}
\seealso{
Other deduplication functions:
\code{\link[=duplicateAllOnes]{duplicateAllOnes()}},
\code{\link[=duplicateAllOnesGroupSize]{duplicateAllOnesGroupSize()}},
\code{\link[=duplicateDecide]{duplicateDecide()}},
\code{\link[=duplicateFields]{duplicateFields()}},
\code{\link[=duplicateFinder]{duplicateFinder()}},
\code{\link[=duplicateFix]{duplicateFix()}},
\code{\link[=duplicateNewest]{duplicateNewest()}},
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=duplicateRecordType]{duplicateRecordType()}},
\code{\link[=duplicateSample]{duplicateSample()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
