% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateNewest}
\alias{duplicateNewest}
\title{Find the most recent records out of a group of duplicates}
\usage{
duplicateNewest(duplicates)
}
\arguments{
\item{duplicates}{The tibble created by \code{\link{duplicateID}}}
}
\description{
The internal \code{duplicateNewest} function is used inside of
\code{\link{duplicateFix}} to filter groups of duplicates to the most recent
records out of each group.
}
\seealso{
Other deduplication functions:
\code{\link[=duplicateAllOnes]{duplicateAllOnes()}},
\code{\link[=duplicateAllOnesGroupSize]{duplicateAllOnesGroupSize()}},
\code{\link[=duplicateDecide]{duplicateDecide()}},
\code{\link[=duplicateFields]{duplicateFields()}},
\code{\link[=duplicateFinder]{duplicateFinder()}},
\code{\link[=duplicateFix]{duplicateFix()}},
\code{\link[=duplicateID]{duplicateID()}},
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=duplicateRecordType]{duplicateRecordType()}},
\code{\link[=duplicateSample]{duplicateSample()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
