% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicatePlot}
\alias{duplicatePlot}
\title{Plot duplicates}
\usage{
duplicatePlot(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Plot which fields are duplicates of individual hunters.
}
\seealso{
Other deduplication functions:
\code{\link[=duplicateAllOnes]{duplicateAllOnes()}},
\code{\link[=duplicateAllOnesGroupSize]{duplicateAllOnesGroupSize()}},
\code{\link[=duplicateDecide]{duplicateDecide()}},
\code{\link[=duplicateFields]{duplicateFields()}},
\code{\link[=duplicateFinder]{duplicateFinder()}},
\code{\link[=duplicateFix]{duplicateFix()}},
\code{\link[=duplicateID]{duplicateID()}},
\code{\link[=duplicateNewest]{duplicateNewest()}},
\code{\link[=duplicateRecordType]{duplicateRecordType()}},
\code{\link[=duplicateSample]{duplicateSample()}}

Other plotting functions:
\code{\link[=errorLevelErrorsByField]{errorLevelErrorsByField()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotDL]{errorPlotDL()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorPlotStates]{errorPlotStates()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=redFlags]{redFlags()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
\concept{plotting functions}
