% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encodingCheck.R
\name{encodingCheck}
\alias{encodingCheck}
\title{Check if data are UTF-8}
\usage{
encodingCheck(path)
}
\arguments{
\item{path}{The path to the data .txt files}
}
\description{
Check the encoding of all files in download directory
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
