% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorLevelErrorsByField}
\alias{errorLevelErrorsByField}
\title{Calculate error-level errors by field}
\usage{
errorLevelErrorsByField(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}} or \code{\link{correct}}}
}
\description{
The internal \code{errorLevelErrorsByField} function calculates a summary
table of the count of errors and proportion of erroneous values by field.
}
\seealso{
Other error-finding functions:
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotDL]{errorPlotDL()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorPlotStates]{errorPlotStates()}},
\code{\link[=errorTable]{errorTable()}},
\code{\link[=errorTableSummary]{errorTableSummary()}},
\code{\link[=pullErrors]{pullErrors()}},
\code{\link[=redFlags]{redFlags()}}

Other plotting functions:
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotDL]{errorPlotDL()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorPlotStates]{errorPlotStates()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=redFlags]{redFlags()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{error-finding functions}
\concept{plotting functions}
