% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlot_states.R
\name{errorPlot_states}
\alias{errorPlot_states}
\title{Plot errors by state}
\usage{
errorPlot_states(x, type)
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{type}{Type of plot to create
Acceptable values include:
\itemize{
\item proportion - Number of errors divided by the number of records per state
\item count - Count of total errors per state
}}
}
\description{
Create a bar plot of errors by state, either by count or proportion
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
