% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorPlotDL}
\alias{errorPlotDL}
\title{Plot errors across download cycles}
\usage{
errorPlotDL(proofed_data, loc = "all")
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}} or \code{\link{correct}}}

\item{loc}{The location that errors should be plotted for. Acceptable values
include:
\itemize{
\item "all" - all states
\item a two-letter abbreviation for a US state; one of:
\itemize{
\item AL, AK, AZ, AR, CA, CO, CT, DE, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY}
}}
}
\description{
Create a plot of errors per download cycle, either by all states in the data
set or a specific state, province, or territory.
}
\seealso{
Other error-finding functions:
\code{\link[=errorLevelErrorsByField]{errorLevelErrorsByField()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorPlotStates]{errorPlotStates()}},
\code{\link[=errorTable]{errorTable()}},
\code{\link[=errorTableSummary]{errorTableSummary()}},
\code{\link[=pullErrors]{pullErrors()}},
\code{\link[=redFlags]{redFlags()}}

Other plotting functions:
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=errorLevelErrorsByField]{errorLevelErrorsByField()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorPlotStates]{errorPlotStates()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=redFlags]{redFlags()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{error-finding functions}
\concept{plotting functions}
