% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorPlotStates}
\alias{errorPlotStates}
\title{Plot errors by state}
\usage{
errorPlotStates(proofed_data, threshold = NA)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}} or \code{\link{correct}}}

\item{threshold}{Optional. A decimal value above which error proportions
should be plotted.}
}
\description{
Create a bar plot of errors by state, either by count or proportion.
}
\seealso{
Other error-finding functions:
\code{\link[=errorLevelErrorsByField]{errorLevelErrorsByField()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotDL]{errorPlotDL()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorTable]{errorTable()}},
\code{\link[=errorTableSummary]{errorTableSummary()}},
\code{\link[=pullErrors]{pullErrors()}},
\code{\link[=redFlags]{redFlags()}}

Other plotting functions:
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=errorLevelErrorsByField]{errorLevelErrorsByField()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotDL]{errorPlotDL()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=redFlags]{redFlags()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{error-finding functions}
\concept{plotting functions}
