% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failCR}
\alias{failCR}
\title{Fail if crane permit file contains bad values}
\usage{
failCR(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
Internal function that applies to crane permit files.
}
\seealso{
Other failure functions:
\code{\link[=failBTPI]{failBTPI()}},
\code{\link[=failDLdate]{failDLdate()}},
\code{\link[=failDLstate]{failDLstate()}},
\code{\link[=failProofed]{failProofed()}},
\code{\link[=failState]{failState()}},
\code{\link[=failTF]{failTF()}},
\code{\link[=failWidths]{failWidths()}},
\code{\link[=failYear]{failYear()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
