% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failDLstate}
\alias{failDLstate}
\title{Fail if there is an NA in dl_state}
\usage{
failDLstate(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
The internal \code{failDLstate} function is used inside of
\code{\link{write_hip}}.
}
\seealso{
Other failure functions:
\code{\link[=failBTPI]{failBTPI()}},
\code{\link[=failCR]{failCR()}},
\code{\link[=failDLdate]{failDLdate()}},
\code{\link[=failProofed]{failProofed()}},
\code{\link[=failState]{failState()}},
\code{\link[=failTF]{failTF()}},
\code{\link[=failWidths]{failWidths()}},
\code{\link[=failYear]{failYear()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
