% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failField}
\alias{failField}
\title{Fail if bad field supplied}
\usage{
failField(field)
}
\arguments{
\item{field}{Any field name in HIP tibble; should be one of:
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_yr, email}}
}
\description{
Internal function that fails if a bad field name is provided.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
