% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failLocation}
\alias{failLocation}
\title{Fail if bad location supplied}
\usage{
failLocation(loc)
}
\arguments{
\item{loc}{Must be one of:
\itemize{
\item "all"
\item "none"
\item a two-letter abbreviation for a US state; one of:
\itemize{
\item AL, AK, AZ, AR, CA, CO, CT, DE, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY}
}}
}
\description{
Internal function that fails if something other than a state abbreviation,
"all", or "none" is provided.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
