% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failWidths}
\alias{failWidths}
\title{Fail if field widths are exceeded by any value}
\usage{
failWidths(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
Internal function that fails inside of \code{\link{write_hip}} if any value
exceeds designated field widths.
}
\seealso{
Other failure functions:
\code{\link[=failBTPI]{failBTPI()}},
\code{\link[=failCR]{failCR()}},
\code{\link[=failDLdate]{failDLdate()}},
\code{\link[=failDLstate]{failDLstate()}},
\code{\link[=failProofed]{failProofed()}},
\code{\link[=failState]{failState()}},
\code{\link[=failTF]{failTF()}},
\code{\link[=failYear]{failYear()}}

Other writing functions:
\code{\link[=writeReport]{writeReport()}},
\code{\link[=write_hip]{write_hip()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
\concept{writing functions}
