% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{fileRename}
\alias{fileRename}
\title{Raw HIP file name correction}
\usage{
fileRename(path, year)
}
\arguments{
\item{path}{Directory to download folder containing new HIP files}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
This function overwrites HIP filenames. Files in the supplied directory are
renamed by converting the Julian date to YYYYMMDD format. State abbreviations
that are in lowercase format are capitalized.
}
\seealso{
Other file functions:
\code{\link[=fileCheck]{fileCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{file functions}
