% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{filterOutOregonPermits}
\alias{filterOutOregonPermits}
\title{Filter out Oregon in-line permit records}
\usage{
filterOutOregonPermits(deduplicated_data)
}
\arguments{
\item{deduplicated_data}{The object created after deduplicating data with \code{\link{duplicateFix}}}
}
\description{
The internal \code{filterOutOregonPermits} function removes Oregon in-line permits inside of \code{\link{bagCheck}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
