% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDuplicates.R
\name{findDuplicates}
\alias{findDuplicates}
\title{Find duplicates}
\usage{
findDuplicates(x, return = "plot")
}
\arguments{
\item{x}{A cleaned data table created by \code{\link{clean}}}

\item{return}{Option to return either a plot or a table; default value is "plot"}
}
\description{
Determine how many duplicate records are in the data. Plot and tabulate which fields are duplicates of individual hunters (i.e. data grouped by first name, last name, city, state, and birth date).
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
