% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDuplicates.R
\name{fixDuplicates}
\alias{fixDuplicates}
\title{Fix duplicates}
\usage{
fixDuplicates(x)
}
\arguments{
\item{x}{A cleaned data table created by \code{\link{clean}}}
}
\description{
Consolidate duplicate records that stem from HIP data and special permit
information being stored in separate rows. If other duplicates exist not due
to this specific reason, move the records to a separate table. Records in the
consolidated table will indicate they are combined HIP/Permit records (value
"hip-permit") in a new column labeled "record_type". All other original
non-duplicated records will be indicated in the "record_type" field as being
HIP only (value "hip").
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
