% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{fixMiddleInitials}
\alias{fixMiddleInitials}
\title{Fix middle initials}
\usage{
fixMiddleInitials(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
The internal \code{fixMiddleInitials} function changes non-alphabetic characters in the middle initial column to NA.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
