% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{formatZip}
\alias{formatZip}
\title{Format zip codes}
\usage{
formatZip(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{formatZip} function formats zip codes.
}
\seealso{
Other cleaning functions:
\code{\link[=btpiPermitBagFix]{btpiPermitBagFix()}},
\code{\link[=clean]{clean()}},
\code{\link[=cranePermitBagFix]{cranePermitBagFix()}},
\code{\link[=inLinePermitDNHFix]{inLinePermitDNHFix()}},
\code{\link[=missingPIIFilter]{missingPIIFilter()}},
\code{\link[=moveSuffixes]{moveSuffixes()}},
\code{\link[=namesToUppercase]{namesToUppercase()}},
\code{\link[=zipCheck]{zipCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{cleaning functions}
