% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadAddress}
\alias{getBadAddress}
\title{Get bad address values}
\usage{
getBadAddress(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadAddress} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{address} values that are not expected. Used by \code{\link{qAddress}}.
}
\seealso{
Other proofing functions:
\code{\link[=getBadBirthDate]{getBadBirthDate()}},
\code{\link[=getBadCity]{getBadCity()}},
\code{\link[=getBadFirstName]{getBadFirstName()}},
\code{\link[=getBadHuntMigBirds]{getBadHuntMigBirds()}},
\code{\link[=getBadLastName]{getBadLastName()}},
\code{\link[=getBadMiddle]{getBadMiddle()}},
\code{\link[=getBadRegYear]{getBadRegYear()}},
\code{\link[=getBadState]{getBadState()}},
\code{\link[=getBadSuffix]{getBadSuffix()}},
\code{\link[=getBadTitle]{getBadTitle()}},
\code{\link[=getBadZIP]{getBadZIP()}},
\code{\link[=proof]{proof()}},
\code{\link[=proofBadEmails]{proofBadEmails()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
