% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadFirstName}
\alias{getBadFirstName}
\title{Get bad first name values}
\usage{
getBadFirstName(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadFirstName} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{firstname} values that are not expected. Used by
\code{\link{qFirstName}}.
}
\seealso{
Other proofing functions:
\code{\link[=getBadAddress]{getBadAddress()}},
\code{\link[=getBadBirthDate]{getBadBirthDate()}},
\code{\link[=getBadCity]{getBadCity()}},
\code{\link[=getBadHuntMigBirds]{getBadHuntMigBirds()}},
\code{\link[=getBadLastName]{getBadLastName()}},
\code{\link[=getBadMiddle]{getBadMiddle()}},
\code{\link[=getBadRegYear]{getBadRegYear()}},
\code{\link[=getBadState]{getBadState()}},
\code{\link[=getBadSuffix]{getBadSuffix()}},
\code{\link[=getBadTitle]{getBadTitle()}},
\code{\link[=getBadZIP]{getBadZIP()}},
\code{\link[=proof]{proof()}},
\code{\link[=proofBadEmails]{proofBadEmails()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
