% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadHuntMigBirds}
\alias{getBadHuntMigBirds}
\title{Get bad hunt mig birds values}
\usage{
getBadHuntMigBirds(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadHuntMigBirds} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{hunt_mig_birds} values that are not expected.
}
\seealso{
Other proofing functions:
\code{\link[=getBadAddress]{getBadAddress()}},
\code{\link[=getBadBirthDate]{getBadBirthDate()}},
\code{\link[=getBadCity]{getBadCity()}},
\code{\link[=getBadFirstName]{getBadFirstName()}},
\code{\link[=getBadLastName]{getBadLastName()}},
\code{\link[=getBadMiddle]{getBadMiddle()}},
\code{\link[=getBadRegYear]{getBadRegYear()}},
\code{\link[=getBadState]{getBadState()}},
\code{\link[=getBadSuffix]{getBadSuffix()}},
\code{\link[=getBadTitle]{getBadTitle()}},
\code{\link[=getBadZIP]{getBadZIP()}},
\code{\link[=proof]{proof()}},
\code{\link[=proofBadEmails]{proofBadEmails()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
