% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hunterSample.R
\name{hunterSample}
\alias{hunterSample}
\title{Sample hunters}
\usage{
hunterSample(x, random = NA)
}
\arguments{
\item{x}{The object created after finalizing data with \code{\link{correct}}}

\item{random}{Optional ability to supply the sample size desired as an integer. The sample will be subset from the data at random and is not reproducible.}
}
\description{
After finalizing the data with \code{\link{correct}}, select a subset of hunters with valid registrations and addresses.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
