% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{huntMigBirdsMessage}
\alias{huntMigBirdsMessage}
\title{Return message if records with bad values for hunt_mig_birds are detected}
\usage{
huntMigBirdsMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{huntMigBirdsMessage} function is used inside of
\code{\link{readMessages}}
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
