% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{idBlankFiles}
\alias{idBlankFiles}
\title{Identify blank files}
\usage{
idBlankFiles(filelist)
}
\arguments{
\item{filelist}{The file list tibble created by \code{\link{listFiles}}}
}
\description{
The internal \code{idBlankFiles} function is used inside of
\code{\link{read_hip}} to identify files in the list that contain no data.
}
\seealso{
Other reading functions:
\code{\link[=checkFileNameDateFormat]{checkFileNameDateFormat()}},
\code{\link[=checkFileNameStateAbbr]{checkFileNameStateAbbr()}},
\code{\link[=dropBlankFiles]{dropBlankFiles()}},
\code{\link[=dropBlankLines]{dropBlankLines()}},
\code{\link[=ignoreHolds]{ignoreHolds()}},
\code{\link[=ignoreLifetime]{ignoreLifetime()}},
\code{\link[=ignorePermits]{ignorePermits()}},
\code{\link[=listFiles]{listFiles()}},
\code{\link[=readTimeMessage]{readTimeMessage()}},
\code{\link[=read_hip]{read_hip()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{reading functions}
