% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identicalBags.R
\name{identicalBags}
\alias{identicalBags}
\title{Check for identical bag columns}
\usage{
identicalBags(x)
}
\arguments{
\item{x}{A HIP data tibble}
}
\description{
Check to make sure the data from each file don't have any columns that are exactly the same.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
