% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{inLinePermitDNHFix}
\alias{inLinePermitDNHFix}
\title{In-line permit did-not-hunt fix}
\usage{
inLinePermitDNHFix(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{inLinePermitDNHFix} function changes any presumed solo
permit from OR or WA indicating "did not hunt" in the hunt_mig_birds field if
one or more of the band_tailed_pigeon, brant, or seaducks fields indicate
hunting.
}
\seealso{
Other cleaning functions:
\code{\link[=btpiPermitBagFix]{btpiPermitBagFix()}},
\code{\link[=clean]{clean()}},
\code{\link[=cranePermitBagFix]{cranePermitBagFix()}},
\code{\link[=formatZip]{formatZip()}},
\code{\link[=missingPIIFilter]{missingPIIFilter()}},
\code{\link[=moveSuffixes]{moveSuffixes()}},
\code{\link[=namesToUppercase]{namesToUppercase()}},
\code{\link[=zipCheck]{zipCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{cleaning functions}
