% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{interStateDuplicatesMessage}
\alias{interStateDuplicatesMessage}
\title{Inter-state duplicates message}
\usage{
interStateDuplicatesMessage(raw_data, n_threshold = 100, p_threshold = 0.05)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}

\item{n_threshold}{A whole number; the threshold below which inter-state
duplicates will not be reported, unless \code{p_threshold} is exceeded.
Defaults to n = 100 inter-state duplicates.}

\item{p_threshold}{A number ranging from 0 to 1; the threshold below which
inter-state duplicates will not be reported, unless \code{n_threshold} is
exceeded. Defaults to 0.05, or inter-state duplicates exceeding more than
5\% of the number of submitted registrations by the first download state OR
the second download state.}
}
\description{
The internal \code{interStateDuplicatesMessage} function returns a message
for duplicates found among files submitted from different states.
}
\seealso{
Other quality functions:
\code{\link[=inLinePermitDNHMessage]{inLinePermitDNHMessage()}},
\code{\link[=missingEmailsMessage]{missingEmailsMessage()}},
\code{\link[=missingPIIMessage]{missingPIIMessage()}},
\code{\link[=naBagsMessage]{naBagsMessage()}},
\code{\link[=nonDigitBagsMessage]{nonDigitBagsMessage()}},
\code{\link[=nonResidentMessage]{nonResidentMessage()}},
\code{\link[=permitFileBagsMessage]{permitFileBagsMessage()}},
\code{\link[=qAddress]{qAddress()}},
\code{\link[=qBags]{qBags()}},
\code{\link[=qBirthDate]{qBirthDate()}},
\code{\link[=qBirthDateRange]{qBirthDateRange()}},
\code{\link[=qCity]{qCity()}},
\code{\link[=qFirstName]{qFirstName()}},
\code{\link[=qHuntMigBirds]{qHuntMigBirds()}},
\code{\link[=qIssueDateRange]{qIssueDateRange()}},
\code{\link[=qLastName]{qLastName()}},
\code{\link[=qMiddle]{qMiddle()}},
\code{\link[=qRegistrationYear]{qRegistrationYear()}},
\code{\link[=qState]{qState()}},
\code{\link[=qSuffix]{qSuffix()}},
\code{\link[=qSummary]{qSummary()}},
\code{\link[=qTitle]{qTitle()}},
\code{\link[=qZIP]{qZIP()}},
\code{\link[=qualityMessages]{qualityMessages()}},
\code{\link[=testRecordMessage]{testRecordMessage()}},
\code{\link[=zeroBagsMessage]{zeroBagsMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
