% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issueAssign}
\alias{issueAssign}
\title{Assign registration_yr values using issueDecide}
\usage{
issueAssign(clean_data, year)
}
\arguments{
\item{clean_data}{The object created after cleaning data with
\code{\link{clean}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
The internal \code{issueAssign} function is used inside of
\code{\link{issueCheck}} to assign a new registration year to each record
depending on the outcome of \code{\link{issueDecide}}.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueCheck]{issueCheck()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issueMessages]{issueMessages()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=issuePrint]{issuePrint()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=regYearEditMessage]{regYearEditMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=twoSeasonMessage]{twoSeasonMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
