% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issueCheck}
\alias{issueCheck}
\title{Check issue dates and license years}
\usage{
issueCheck(clean_data, year, plot = FALSE)
}
\arguments{
\item{clean_data}{The object created after cleaning data with
\code{\link{clean}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}

\item{plot}{Create a plot? Default is FALSE}
}
\description{
After cleaning the data with \code{\link{clean}}, ensure each record is
assigned the appropriate \code{registration_yr}.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueAssign]{issueAssign()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issueMessages]{issueMessages()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=issuePrint]{issuePrint()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=regYearEditMessage]{regYearEditMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=twoSeasonMessage]{twoSeasonMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
