% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issueMessages}
\alias{issueMessages}
\title{Return messages to console for issueCheck insights}
\usage{
issueMessages(clean_data, issue_assignments)
}
\arguments{
\item{clean_data}{The product of \code{\link{clean}}}

\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{issueMessages} function is used inside of
\code{\link{issueCheck}} to return messages for bad issue_date values, count
of past registrations, count of future registrations, and 2-season overlap
registrations.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueAssign]{issueAssign()}},
\code{\link[=issueCheck]{issueCheck()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=issuePrint]{issuePrint()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=regYearEditMessage]{regYearEditMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=twoSeasonMessage]{twoSeasonMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
