% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePlot}
\alias{issuePlot}
\title{Plot issue date errors}
\usage{
issuePlot(issue_assignments, year)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
The internal \code{issuePlot} function plots bad \code{issue_date} values.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueAssign]{issueAssign()}},
\code{\link[=issueCheck]{issueCheck()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issueMessages]{issueMessages()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=issuePrint]{issuePrint()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=regYearEditMessage]{regYearEditMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=twoSeasonMessage]{twoSeasonMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}

Other plotting functions:
\code{\link[=duplicatePlot]{duplicatePlot()}},
\code{\link[=errorLevelErrorsByField]{errorLevelErrorsByField()}},
\code{\link[=errorLevelErrorsByState]{errorLevelErrorsByState()}},
\code{\link[=errorPlotDL]{errorPlotDL()}},
\code{\link[=errorPlotFields]{errorPlotFields()}},
\code{\link[=errorPlotStates]{errorPlotStates()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=redFlags]{redFlags()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
\concept{plotting functions}
