% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePrint}
\alias{issuePrint}
\title{Print results of issueCheck changes}
\usage{
issuePrint(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{issuePrint} function is used inside of
\code{\link{issueCheck}} to print the number of registration types (past,
future, current, etc) per download state and registration year.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueAssign]{issueAssign()}},
\code{\link[=issueCheck]{issueCheck()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issueMessages]{issueMessages()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=regYearEditMessage]{regYearEditMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=twoSeasonMessage]{twoSeasonMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
