% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manualFix.R
\name{manualFix}
\alias{manualFix}
\title{Manually correct data}
\usage{
manualFix(x, field, error_value, correct_value)
}
\arguments{
\item{x}{A corrected data table created by \code{\link{correct}}}

\item{field}{Field that contains value(s) to be corrected
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_year, email}}

\item{error_value}{Verbatim cell contents of value to be corrected}

\item{correct_value}{Desired value to replace error_value}
}
\description{
Replace incorrect cell contents with a desired value.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
