% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{missingPIIFilter}
\alias{missingPIIFilter}
\title{Missing PII filter}
\usage{
missingPIIFilter(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{missingPIIFilter} function filters out HIP registrations
that are missing critical pieces of contact information.
}
\seealso{
Other cleaning functions:
\code{\link[=btpiPermitBagFix]{btpiPermitBagFix()}},
\code{\link[=clean]{clean()}},
\code{\link[=cranePermitBagFix]{cranePermitBagFix()}},
\code{\link[=formatZip]{formatZip()}},
\code{\link[=inLinePermitDNHFix]{inLinePermitDNHFix()}},
\code{\link[=moveSuffixes]{moveSuffixes()}},
\code{\link[=namesToUppercase]{namesToUppercase()}},
\code{\link[=zipCheck]{zipCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{cleaning functions}
