% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{moveSuffixes}
\alias{moveSuffixes}
\title{Move suffixes}
\usage{
moveSuffixes(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{moveSuffixes} function moves suffixes from first name or
last name columns into the suffix column and performs other cleaning steps.
This function catches values from 1 to 20 in Roman numerals and numeric,
excluding XVIII since the database limit is 4 characters.
}
\seealso{
Other cleaning functions:
\code{\link[=btpiPermitBagFix]{btpiPermitBagFix()}},
\code{\link[=clean]{clean()}},
\code{\link[=cranePermitBagFix]{cranePermitBagFix()}},
\code{\link[=formatZip]{formatZip()}},
\code{\link[=inLinePermitDNHFix]{inLinePermitDNHFix()}},
\code{\link[=missingPIIFilter]{missingPIIFilter()}},
\code{\link[=namesToUppercase]{namesToUppercase()}},
\code{\link[=zipCheck]{zipCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{cleaning functions}
