% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{namesToUppercase}
\alias{namesToUppercase}
\title{Names to uppercase}
\usage{
namesToUppercase(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{namesToUppercase} function converts name elements to
uppercase for easier string cleaning.
}
\seealso{
Other cleaning functions:
\code{\link[=btpiPermitBagFix]{btpiPermitBagFix()}},
\code{\link[=clean]{clean()}},
\code{\link[=cranePermitBagFix]{cranePermitBagFix()}},
\code{\link[=formatZip]{formatZip()}},
\code{\link[=inLinePermitDNHFix]{inLinePermitDNHFix()}},
\code{\link[=missingPIIFilter]{missingPIIFilter()}},
\code{\link[=moveSuffixes]{moveSuffixes()}},
\code{\link[=zipCheck]{zipCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{cleaning functions}
