% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDropped.R
\name{nDroppedClean}
\alias{nDroppedClean}
\title{Summary table of how many registrations were dropped by clean}
\usage{
nDroppedClean(raw_data, clean_data, by_state = FALSE)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}

\item{clean_data}{The object created after cleaning data with
\code{\link{clean}}}

\item{by_state}{TRUE summarizes the number of dropped records by state; FALSE
(default) does not summarize}
}
\description{
Internal function that helps \code{\link{nDropped}} by summarizing the number
of dropped registrations by \code{\link{clean}}.
}
\seealso{
Other summary functions:
\code{\link[=nDropped]{nDropped()}},
\code{\link[=nDroppedCurrent]{nDroppedCurrent()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{summary functions}
