% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{nonResidentMessage}
\alias{nonResidentMessage}
\title{High non-resident hunter proportion message}
\usage{
nonResidentMessage(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{nonResidentMessage} function returns a message for files
with 10\% or more of \code{state} values that do not match \code{dl_state}.
}
\seealso{
Other quality functions:
\code{\link[=inLinePermitDNHMessage]{inLinePermitDNHMessage()}},
\code{\link[=interStateDuplicatesMessage]{interStateDuplicatesMessage()}},
\code{\link[=missingEmailsMessage]{missingEmailsMessage()}},
\code{\link[=missingPIIMessage]{missingPIIMessage()}},
\code{\link[=naBagsMessage]{naBagsMessage()}},
\code{\link[=nonDigitBagsMessage]{nonDigitBagsMessage()}},
\code{\link[=permitFileBagsMessage]{permitFileBagsMessage()}},
\code{\link[=qAddress]{qAddress()}},
\code{\link[=qBags]{qBags()}},
\code{\link[=qBirthDate]{qBirthDate()}},
\code{\link[=qBirthDateRange]{qBirthDateRange()}},
\code{\link[=qCity]{qCity()}},
\code{\link[=qFirstName]{qFirstName()}},
\code{\link[=qHuntMigBirds]{qHuntMigBirds()}},
\code{\link[=qIssueDateRange]{qIssueDateRange()}},
\code{\link[=qLastName]{qLastName()}},
\code{\link[=qMiddle]{qMiddle()}},
\code{\link[=qRegistrationYear]{qRegistrationYear()}},
\code{\link[=qState]{qState()}},
\code{\link[=qSuffix]{qSuffix()}},
\code{\link[=qSummary]{qSummary()}},
\code{\link[=qTitle]{qTitle()}},
\code{\link[=qZIP]{qZIP()}},
\code{\link[=qualityMessages]{qualityMessages()}},
\code{\link[=testRecordMessage]{testRecordMessage()}},
\code{\link[=zeroBagsMessage]{zeroBagsMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
