% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfStateHunters.R
\name{outOfStateHunters}
\alias{outOfStateHunters}
\title{Find out-of-state hunters}
\usage{
outOfStateHunters(x, type)
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{type}{Type of plot to create
Acceptable values include:
\itemize{
\item proportion - Number of out-of-state hunters divided by the total hunters per state
\item count - Count of out-of-state hunters per state
}}
}
\description{
Create a tibble and plot of hunters who have reported hunting migratory waterfowl outside of the state associated with their address.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
