% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findDuplicates.R
\name{plotDuplicates}
\alias{plotDuplicates}
\title{Plot duplicates}
\usage{
plotDuplicates(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with \code{\link{issueCheck}}}
}
\description{
Plot which fields are duplicates of individual hunters.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
