% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proof}
\alias{proof}
\title{Proof data and flag errors}
\usage{
proof(deduplicated_data, year)
}
\arguments{
\item{deduplicated_data}{The object created after deduplicating data with
\code{\link{duplicateFix}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
After data are deduplicated with \code{\link{duplicateFix}}, asses each
field's values and flag records with unexpected or erroneous values in a new
\code{errors} field.
}
\seealso{
Other proofing functions:
\code{\link[=getBadAddress]{getBadAddress()}},
\code{\link[=getBadBirthDate]{getBadBirthDate()}},
\code{\link[=getBadCity]{getBadCity()}},
\code{\link[=getBadFirstName]{getBadFirstName()}},
\code{\link[=getBadHuntMigBirds]{getBadHuntMigBirds()}},
\code{\link[=getBadLastName]{getBadLastName()}},
\code{\link[=getBadMiddle]{getBadMiddle()}},
\code{\link[=getBadRegYear]{getBadRegYear()}},
\code{\link[=getBadState]{getBadState()}},
\code{\link[=getBadSuffix]{getBadSuffix()}},
\code{\link[=getBadTitle]{getBadTitle()}},
\code{\link[=getBadZIP]{getBadZIP()}},
\code{\link[=proofBadEmails]{proofBadEmails()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
