% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proofBadEmails}
\alias{proofBadEmails}
\title{Proof bad emails}
\usage{
proofBadEmails(keyed_data)
}
\arguments{
\item{keyed_data}{An tibble used internally in \code{\link{proof}}}
}
\description{
The internal \code{proofBadEmails} function is used inside of
\code{\link{proof}} to find poorly formatted or intentionally obfuscative
email addresses. Email addresses that match \code{REGEX_EMAIL} or are
\code{NA} are considered good emails.
}
\seealso{
Other proofing functions:
\code{\link[=getBadAddress]{getBadAddress()}},
\code{\link[=getBadBirthDate]{getBadBirthDate()}},
\code{\link[=getBadCity]{getBadCity()}},
\code{\link[=getBadFirstName]{getBadFirstName()}},
\code{\link[=getBadHuntMigBirds]{getBadHuntMigBirds()}},
\code{\link[=getBadLastName]{getBadLastName()}},
\code{\link[=getBadMiddle]{getBadMiddle()}},
\code{\link[=getBadRegYear]{getBadRegYear()}},
\code{\link[=getBadState]{getBadState()}},
\code{\link[=getBadSuffix]{getBadSuffix()}},
\code{\link[=getBadTitle]{getBadTitle()}},
\code{\link[=getBadZIP]{getBadZIP()}},
\code{\link[=proof]{proof()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
