% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qBirthDateRange}
\alias{qBirthDateRange}
\title{Birth date range message}
\usage{
qBirthDateRange(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{qBirthDateRange} function returns a message for
\code{birth_date} values in a file that do not span a period of time more
than 365 days.
}
\seealso{
Other quality functions:
\code{\link[=inLinePermitDNHMessage]{inLinePermitDNHMessage()}},
\code{\link[=interStateDuplicatesMessage]{interStateDuplicatesMessage()}},
\code{\link[=missingEmailsMessage]{missingEmailsMessage()}},
\code{\link[=missingPIIMessage]{missingPIIMessage()}},
\code{\link[=naBagsMessage]{naBagsMessage()}},
\code{\link[=nonDigitBagsMessage]{nonDigitBagsMessage()}},
\code{\link[=nonResidentMessage]{nonResidentMessage()}},
\code{\link[=permitFileBagsMessage]{permitFileBagsMessage()}},
\code{\link[=qAddress]{qAddress()}},
\code{\link[=qBags]{qBags()}},
\code{\link[=qBirthDate]{qBirthDate()}},
\code{\link[=qCity]{qCity()}},
\code{\link[=qFirstName]{qFirstName()}},
\code{\link[=qHuntMigBirds]{qHuntMigBirds()}},
\code{\link[=qIssueDateRange]{qIssueDateRange()}},
\code{\link[=qLastName]{qLastName()}},
\code{\link[=qMiddle]{qMiddle()}},
\code{\link[=qRegistrationYear]{qRegistrationYear()}},
\code{\link[=qState]{qState()}},
\code{\link[=qSuffix]{qSuffix()}},
\code{\link[=qSummary]{qSummary()}},
\code{\link[=qTitle]{qTitle()}},
\code{\link[=qZIP]{qZIP()}},
\code{\link[=qualityMessages]{qualityMessages()}},
\code{\link[=testRecordMessage]{testRecordMessage()}},
\code{\link[=zeroBagsMessage]{zeroBagsMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
