% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qDLstate}
\alias{qDLstate}
\title{Return message if there is an NA in dl_state}
\usage{
qDLstate(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qDLstate} function is used inside of
\code{\link{qualityMessages}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
