% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{qualityCheck}
\alias{qualityCheck}
\title{Run checks for common or catastrophic data quality issues}
\usage{
qualityCheck(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The \code{qualityCheck} function reports data quality issues to the user.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
