% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qualityMessages}
\alias{qualityMessages}
\title{Return messages to console for common or catastrophic data quality issues}
\usage{
qualityMessages(raw_data, year)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The \code{qualityMessages} function is used to return messages for
missing PII, missing email addresses, all-zero bag records, non-numeric bag
values, NAs in dl_state, NAs in dl_date, bad title assignments, high
proportions of non-resident hunters in a file, inter-state duplicates, low
range of issue dates in a file, and low range of birth dates in a file.
}
\seealso{
Other quality functions:
\code{\link[=inLinePermitDNHMessage]{inLinePermitDNHMessage()}},
\code{\link[=interStateDuplicatesMessage]{interStateDuplicatesMessage()}},
\code{\link[=missingEmailsMessage]{missingEmailsMessage()}},
\code{\link[=missingPIIMessage]{missingPIIMessage()}},
\code{\link[=naBagsMessage]{naBagsMessage()}},
\code{\link[=nonDigitBagsMessage]{nonDigitBagsMessage()}},
\code{\link[=nonResidentMessage]{nonResidentMessage()}},
\code{\link[=permitFileBagsMessage]{permitFileBagsMessage()}},
\code{\link[=qAddress]{qAddress()}},
\code{\link[=qBags]{qBags()}},
\code{\link[=qBirthDate]{qBirthDate()}},
\code{\link[=qBirthDateRange]{qBirthDateRange()}},
\code{\link[=qCity]{qCity()}},
\code{\link[=qFirstName]{qFirstName()}},
\code{\link[=qHuntMigBirds]{qHuntMigBirds()}},
\code{\link[=qIssueDateRange]{qIssueDateRange()}},
\code{\link[=qLastName]{qLastName()}},
\code{\link[=qMiddle]{qMiddle()}},
\code{\link[=qRegistrationYear]{qRegistrationYear()}},
\code{\link[=qState]{qState()}},
\code{\link[=qSuffix]{qSuffix()}},
\code{\link[=qSummary]{qSummary()}},
\code{\link[=qTitle]{qTitle()}},
\code{\link[=qZIP]{qZIP()}},
\code{\link[=testRecordMessage]{testRecordMessage()}},
\code{\link[=zeroBagsMessage]{zeroBagsMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
