% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questions.R
\name{questionYear}
\alias{questionYear}
\title{Question if correct year supplied}
\usage{
questionYear(year)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}}}
}
\description{
Internal function that questions if the year supplied is intended to be
different than the current season year by returning a message to the console.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
