% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{read_hip}
\alias{read_hip}
\title{Read in data}
\usage{
read_hip(path, unique = TRUE, state = NA, season = FALSE)
}
\arguments{
\item{path}{File path to the folder containing HIP .txt files}

\item{unique}{Return a distinct frame? Defaults to TRUE}

\item{state}{When specified, reads in download data from a specified state.
Must match a two-letter abbreviation for a US state (excluding HI).}

\item{season}{If set as TRUE, selects only folders starting with "DL" in a a
season's upper-level directory}
}
\description{
Compile data from state-exported text files by providing a path to the
download directory.
}
\seealso{
Other reading functions:
\code{\link[=checkFileNameDateFormat]{checkFileNameDateFormat()}},
\code{\link[=checkFileNameStateAbbr]{checkFileNameStateAbbr()}},
\code{\link[=dropBlankFiles]{dropBlankFiles()}},
\code{\link[=dropBlankLines]{dropBlankLines()}},
\code{\link[=idBlankFiles]{idBlankFiles()}},
\code{\link[=ignoreHolds]{ignoreHolds()}},
\code{\link[=ignoreLifetime]{ignoreLifetime()}},
\code{\link[=ignorePermits]{ignorePermits()}},
\code{\link[=listFiles]{listFiles()}},
\code{\link[=readTimeMessage]{readTimeMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{reading functions}
