% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{readMessages}
\alias{readMessages}
\title{Return messages to console for common or catastrophic read_hip issues}
\usage{
readMessages(pulled_data)
}
\arguments{
\item{pulled_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{readMessages} function is used inside of \code{\link{read_hip}} to return messages for missing PII, missing email addresses, all-zero bag records, non-numeric bag values, NAs in dl_state, and NAs in dl_date.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
