% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{readTimeMessage}
\alias{readTimeMessage}
\title{Read time message}
\usage{
readTimeMessage(raw_data, starttime, endtime)
}
\arguments{
\item{raw_data}{Intermediate raw data tibble}

\item{starttime}{Start time}

\item{endtime}{End time}
}
\description{
The internal \code{readTimeMessage} function returns a message to the console
informing the user about how long it took to read in the data with
\code{\link{read_hip}}.
}
\seealso{
Other reading functions:
\code{\link[=checkFileNameDateFormat]{checkFileNameDateFormat()}},
\code{\link[=checkFileNameStateAbbr]{checkFileNameStateAbbr()}},
\code{\link[=dropBlankFiles]{dropBlankFiles()}},
\code{\link[=dropBlankLines]{dropBlankLines()}},
\code{\link[=idBlankFiles]{idBlankFiles()}},
\code{\link[=ignoreHolds]{ignoreHolds()}},
\code{\link[=ignoreLifetime]{ignoreLifetime()}},
\code{\link[=ignorePermits]{ignorePermits()}},
\code{\link[=listFiles]{listFiles()}},
\code{\link[=read_hip]{read_hip()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{reading functions}
