% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{regYearEditMessage}
\alias{regYearEditMessage}
\title{Return message for changed registration_yr values}
\usage{
regYearEditMessage(clean_data, issue_assignments)
}
\arguments{
\item{clean_data}{The product of \code{\link{clean}}}

\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{regYearEditMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
for how many \code{registration_yr} values were changed.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueAssign]{issueAssign()}},
\code{\link[=issueCheck]{issueCheck()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issueMessages]{issueMessages()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=issuePrint]{issuePrint()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=twoSeasonMessage]{twoSeasonMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
