% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameJulian.R
\name{renameJulian}
\alias{renameJulian}
\title{Julian date conversion in raw HIP filenames}
\usage{
renameJulian(path)
}
\arguments{
\item{path}{Directory to HIP files}
}
\description{
This function overwrites HIP filenames. Files in the supplied directory are renamed by converting the Julian date to YYYYMMDD format.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
