% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{shiftCheck}
\alias{shiftCheck}
\title{Check for frame shifts}
\usage{
shiftCheck(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
Find and print any rows that have a line shift error with number of positions
shifted.
}
\seealso{
\code{\link[=glyphCheck]{glyphCheck()}}

Other shifting functions:
\code{\link[=shiftFinder]{shiftFinder()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{shifting functions}
