% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftFix.R
\name{shiftFix}
\alias{shiftFix}
\title{Correct frame shifts}
\usage{
shiftFix(x, record_id, first_col, last_col, n_shift)
}
\arguments{
\item{x}{A cleaned data table created by \code{\link{clean}}}

\item{record_id}{The ID of the line to be adjusted}

\item{first_col}{First (leftmost) column affected by frame shift}

\item{last_col}{Last (rightmost) column affected by frame shift}

\item{n_shift}{Number of characters that the line needs to be adjusted by}
}
\description{
Shift record values by a designated number of characters.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
