% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{singleIssueDateMessage}
\alias{singleIssueDateMessage}
\title{Single issue date message}
\usage{
singleIssueDateMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{singleIssueDateMessage} function returns a message for
files that contain only one unique \code{issue_date} value.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
