% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{special_OregonHuntYCheck}
\alias{special_OregonHuntYCheck}
\title{Oregon HuntY check}
\usage{
special_OregonHuntYCheck(raw_df)
}
\arguments{
\item{raw_df}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
The internal \code{special_OregonHuntYCheck} function changes any registration from Oregon with HuntY == "0" to "2" if one or more of the band_tailed_pigeon, brant, or seaducks fields indicate hunting.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
