% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{sumLines}
\alias{sumLines}
\title{Sum lines of new data}
\usage{
sumLines(path)
}
\arguments{
\item{path}{File path to the download folder containing HIP .txt files}
}
\description{
The internal \code{sumLines} function returns a data frame containing the sums of the number of lines in new download files. It is used inside of \code{\link{read_hip}} to return a message for any file that did not have all its lines read in.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
