% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{summarizeBadBags}
\alias{summarizeBadBags}
\title{Summarize count and proportion of bad bag values}
\usage{
summarizeBadBags(deduplicated_data, bad_bag_values, x)
}
\arguments{
\item{deduplicated_data}{The object created after deduplicating data with
\code{\link{duplicateFix}}}

\item{bad_bag_values}{Bad bag values df}

\item{x}{Row number}
}
\description{
The internal \code{filterOutOregonPermits} function calculates the count and
proportion of a bad bag value inside of \code{\link{bagCheck}}.
}
\seealso{
Other bag functions:
\code{\link[=bagCheck]{bagCheck()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{bag functions}
