% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{testRecordFilter}
\alias{testRecordFilter}
\title{Filter out test records}
\usage{
testRecordFilter(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
The internal \code{testRecordFilter} function filters out records if they contain all-NA or all-zero bag values.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
