% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_utf8.R
\name{to_utf8}
\alias{to_utf8}
\title{Convert data to utf-8}
\usage{
to_utf8(original, encoding)
}
\arguments{
\item{original}{The path to the original file that needs to be converted to utf-8}

\item{encoding}{The existing encoding of the original file}
}
\description{
Convert text files from their original encoding to utf-8. The new file is added to the same directory as the original file.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
