% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{twoSeasonMessage}
\alias{twoSeasonMessage}
\title{Return message if registrations are detected from two-season overlap windows}
\usage{
twoSeasonMessage(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{twoSeasonMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if registrations are detected from two-season states with \code{issue_date}
value(s) that occur during the overlap window.
}
\seealso{
Other issuance functions:
\code{\link[=badDateMessage]{badDateMessage()}},
\code{\link[=futureDateMessage]{futureDateMessage()}},
\code{\link[=invalidDateMessage]{invalidDateMessage()}},
\code{\link[=issueAssign]{issueAssign()}},
\code{\link[=issueCheck]{issueCheck()}},
\code{\link[=issueDecide]{issueDecide()}},
\code{\link[=issueMessages]{issueMessages()}},
\code{\link[=issuePlot]{issuePlot()}},
\code{\link[=issuePlotDateLabel]{issuePlotDateLabel()}},
\code{\link[=issuePrint]{issuePrint()}},
\code{\link[=pastDateMessage]{pastDateMessage()}},
\code{\link[=regYearEditMessage]{regYearEditMessage()}},
\code{\link[=timeTravelMessage]{timeTravelMessage()}},
\code{\link[=zeroDateMessage]{zeroDateMessage()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
