% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate data}
\usage{
validate(x, type, period = NA)
}
\arguments{
\item{x}{The object created after tidying data with \code{\link{tidy}}}

\item{type}{Type of validation to perform. Acceptable values include:
\itemize{
\item vertical - Checks for repetition vertically in species and/or bag fields, grouped by dl_state and dl_date
\item horizontal - Checks for repetition horizontally, across each record
}}

\item{period}{Time period in which to group the data. The function uses dl_date automatically, but either of the following may be supplied:
\itemize{
\item dl_date - Date the HIP data were downloaded
\item dl_cycle - Download cycle that the HIP data belong to
}}
}
\description{
After tidying the data with \code{\link{tidy}}, check to make sure the data don't have any erroneously repeated values.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
