% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeShiny.R
\name{writeShiny}
\alias{writeShiny}
\title{Write an R markdown report with shiny plots from template}
\usage{
writeShiny(path, type, yr)
}
\arguments{
\item{path}{File path to the folder containing HIP .txt files}

\item{type}{Type of report. One of the following options may be supplied:
\itemize{
\item dl_shiny - for a download cycle summary report
\item season_shiny - for a summary of the entire HIP season}}

\item{yr}{The year in which the Harvest Information Program data were collected}
}
\description{
Create documentation with shiny figures that summarizes HIP download data at a download cycle or season total scale.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
