% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/youthHunters.R
\name{youthHunters}
\alias{youthHunters}
\title{Youth hunters}
\usage{
youthHunters(x, year)
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{year}{The year in which the Harvest Information Program data were collected}
}
\description{
Create a tibble and plot of youth hunters (hunters born < 16 years ago) who have reported hunting migratory waterfowl.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
