% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{badDateMessage}
\alias{badDateMessage}
\title{Return message if bad issue dates detected}
\usage{
badDateMessage(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{badDateMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if invalid \code{issue_date} values are detected.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
