% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bags.R
\name{bagCheck}
\alias{bagCheck}
\title{Check bag values}
\usage{
bagCheck(deduplicated_data)
}
\arguments{
\item{deduplicated_data}{The object created after deduplicating data with
\code{\link{duplicateFix}}}
}
\description{
After fixing the data with \code{\link{duplicateFix}}, see if any unexpected
bag values were submitted.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
