% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{checkFileNameDateFormat}
\alias{checkFileNameDateFormat}
\title{Check HIP file name date formatting}
\usage{
checkFileNameDateFormat(file_list_vector)
}
\arguments{
\item{file_list_vector}{A file list vector}
}
\description{
The internal \code{checkFileNameDateFormat} function is used inside of
\code{\link{read_hip}} to return an error message if any file does not have a
date formatted as YYYYMMDD.
}
\seealso{
Other reading functions: 
\code{\link{checkFileNameStateAbbr}()},
\code{\link{dropBlankFiles}()},
\code{\link{dropBlankLines}()},
\code{\link{idBlankFiles}()},
\code{\link{ignoreHolds}()},
\code{\link{ignoreLifetime}()},
\code{\link{ignorePermits}()},
\code{\link{listFiles}()},
\code{\link{readTimeMessage}()},
\code{\link{read_hip}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{reading functions}
