% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{checkFileNameStateAbbr}
\alias{checkFileNameStateAbbr}
\title{Check HIP file name state abbreviations}
\usage{
checkFileNameStateAbbr(file_list_vector)
}
\arguments{
\item{file_list_vector}{A file list vector}
}
\description{
The internal \code{checkFileNameStateAbbr} function is used inside of
\code{\link{read_hip}} to return an error message if any file does not have
an state abbreviation from the expected 49 continental states.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
