% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean data}
\usage{
clean(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
After reading the data with \code{\link{read_hip}}, reformat and clean the
HIP registrations.
}
\seealso{
Other cleaning functions: 
\code{\link{btpiPermitBagFix}()},
\code{\link{cranePermitBagFix}()},
\code{\link{formatZip}()},
\code{\link{inLinePermitDNHFix}()},
\code{\link{missingPIIFilter}()},
\code{\link{moveSuffixes}()},
\code{\link{namesToUppercase}()},
\code{\link{zipCheck}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{cleaning functions}
