% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Tidy data}
\usage{
clean(x, plot)
}
\arguments{
\item{x}{The object created after reading in data with \code{\link{compile}}}

\item{plot}{If TRUE, plots the records and fields that were filtered out of critical identifying information was missing; default is FALSE.}
}
\description{
After reading the data with \code{\link{compile}}, rename the column names and do preliminary cleaning of firstname, lastname, middle initial, and suffixes.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
