% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\title{Clean data}
\usage{
clean(x)
}
\arguments{
\item{x}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
After reading the data with \code{\link{read_hip}}, rename the column names and do preliminary cleaning of firstname, lastname, middle initial, and suffixes.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
