% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.R
\name{correct}
\alias{correct}
\title{Correct data}
\usage{
correct(proofed_data, year)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
After flagging errors in the data with \code{\link{proof}}, attempt
corrections in all fields. Errors that cannot be programmatically corrected
will be reported for manual correction.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
