% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{dropBlankFiles}
\alias{dropBlankFiles}
\title{Drop blank files}
\usage{
dropBlankFiles(filelist)
}
\arguments{
\item{filelist}{The file list tibble created by \code{\link{listFiles}}}
}
\description{
The internal \code{dropBlankFiles} function is used inside of
\code{\link{read_hip}} to return an error message if blank files exist in the
directory, and remove them from the file list so they are not read in.
}
\seealso{
Other reading functions: 
\code{\link{checkFileNameDateFormat}()},
\code{\link{checkFileNameStateAbbr}()},
\code{\link{dropBlankLines}()},
\code{\link{idBlankFiles}()},
\code{\link{ignoreHolds}()},
\code{\link{ignoreLifetime}()},
\code{\link{ignorePermits}()},
\code{\link{listFiles}()},
\code{\link{readTimeMessage}()},
\code{\link{read_hip}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{reading functions}
