% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{dropBlankLines}
\alias{dropBlankLines}
\title{Drop blank lines}
\usage{
dropBlankLines(raw_data)
}
\arguments{
\item{raw_data}{Intermediate raw data tibble}
}
\description{
The internal \code{dropBlankLines} function is used inside of
\code{\link{read_hip}} to filter out any rows with \code{NA} in every field
or a blank line beginning with \code{Result}.
}
\seealso{
Other reading functions: 
\code{\link{checkFileNameDateFormat}()},
\code{\link{checkFileNameStateAbbr}()},
\code{\link{dropBlankFiles}()},
\code{\link{idBlankFiles}()},
\code{\link{ignoreHolds}()},
\code{\link{ignoreLifetime}()},
\code{\link{ignorePermits}()},
\code{\link{listFiles}()},
\code{\link{readTimeMessage}()},
\code{\link{read_hip}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{reading functions}
