% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateAllOnes}
\alias{duplicateAllOnes}
\title{Flag all-one records in a group of duplicates}
\usage{
duplicateAllOnes(duplicates)
}
\arguments{
\item{duplicates}{The tibble created by \code{\link{duplicateID}}}
}
\description{
The internal \code{duplicateAllOnes} function is used inside of \code{\link{duplicateFix}} to evaluate groups of duplicates for records containing "1" for every bag value.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
