% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateAllOnesGroupSize}
\alias{duplicateAllOnesGroupSize}
\title{Evaluate group sizes of all-one/not-all-one records}
\usage{
duplicateAllOnesGroupSize(duplicates)
}
\arguments{
\item{duplicates}{The tibble created by \code{\link{duplicateID}}}
}
\description{
The internal \code{duplicateAllOnesGroupSize} function is used inside of \code{\link{duplicateFix}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
