% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateDecide}
\alias{duplicateDecide}
\title{Decide which duplicate records should be kept or dropped}
\usage{
duplicateDecide(dupes)
}
\arguments{
\item{dupes}{The tibble created by \code{\link{duplicateID}}}
}
\description{
The internal \code{duplicateDecide} function is used inside of
\code{\link{duplicateFix}} to deduplicate intermediate tibbles.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFinder}()},
\code{\link{duplicateFix}()},
\code{\link{duplicateID}()},
\code{\link{duplicateNewest}()},
\code{\link{duplicatePlot}()},
\code{\link{duplicateRecordType}()},
\code{\link{duplicateSample}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
