% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateFields}
\alias{duplicateFields}
\title{Find causes of duplication}
\usage{
duplicateFields(duplicates, fields)
}
\arguments{
\item{duplicates}{The tibble created by \code{\link{duplicateID}}}

\item{fields}{Name of the column to compare values for}
}
\description{
The internal \code{duplicateFields} function is used inside of
\code{\link{duplicateFinder}} to find which fields have different values
among a group of duplicate registrations.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
