% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateFinder}
\alias{duplicateFinder}
\title{Find duplicates}
\usage{
duplicateFinder(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Determine how many duplicate records are in the data and return a table.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateDecide}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFix}()},
\code{\link{duplicateID}()},
\code{\link{duplicateNewest}()},
\code{\link{duplicatePlot}()},
\code{\link{duplicateRecordType}()},
\code{\link{duplicateSample}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
