% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateFix}
\alias{duplicateFix}
\title{Fix duplicates}
\usage{
duplicateFix(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Resolve duplicate HIP records.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateDecide}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFinder}()},
\code{\link{duplicateID}()},
\code{\link{duplicateNewest}()},
\code{\link{duplicatePlot}()},
\code{\link{duplicateRecordType}()},
\code{\link{duplicateSample}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
