% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateFix}
\alias{duplicateFix}
\title{Fix duplicates}
\usage{
duplicateFix(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with \code{\link{issueCheck}}}
}
\description{
Resolve duplicate HIP records.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
