% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateID}
\alias{duplicateID}
\title{Create a tibble of duplicates with an ID column}
\usage{
duplicateID(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with \code{\link{issueCheck}}}
}
\description{
Return a tibble of duplicates with a duplicate ID column identifying each group of records.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
