% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateNewest}
\alias{duplicateNewest}
\title{Find the most recent records out of a group of duplicates}
\usage{
duplicateNewest(duplicates)
}
\arguments{
\item{duplicates}{The tibble created by \code{\link{duplicateID}}}
}
\description{
The internal \code{duplicateNewest} function is used inside of
\code{\link{duplicateFix}} to filter groups of duplicates to the most recent
records out of each group.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateDecide}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFinder}()},
\code{\link{duplicateFix}()},
\code{\link{duplicateID}()},
\code{\link{duplicatePlot}()},
\code{\link{duplicateRecordType}()},
\code{\link{duplicateSample}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
