% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicatePlot}
\alias{duplicatePlot}
\title{Plot duplicates}
\usage{
duplicatePlot(current_data)
}
\arguments{
\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}
}
\description{
Plot which fields are duplicates of individual hunters.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateDecide}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFinder}()},
\code{\link{duplicateFix}()},
\code{\link{duplicateID}()},
\code{\link{duplicateNewest}()},
\code{\link{duplicateRecordType}()},
\code{\link{duplicateSample}()}

Other plotting functions: 
\code{\link{errorLevelErrorsByField}()},
\code{\link{errorLevelErrorsByState}()},
\code{\link{errorPlotDL}()},
\code{\link{errorPlotFields}()},
\code{\link{errorPlotStates}()},
\code{\link{issuePlot}()},
\code{\link{issuePlotDateLabel}()},
\code{\link{redFlags}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
\concept{plotting functions}
