% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateRecordType}
\alias{duplicateRecordType}
\title{Set record type}
\usage{
duplicateRecordType(duplicates)
}
\arguments{
\item{duplicates}{The tibble created by \code{\link{duplicateID}}}
}
\description{
The internal \code{duplicateRecordType} function is used inside of
\code{\link{duplicateFix}} to set record type of registrations based on each
record's bag values.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateDecide}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFinder}()},
\code{\link{duplicateFix}()},
\code{\link{duplicateID}()},
\code{\link{duplicateNewest}()},
\code{\link{duplicatePlot}()},
\code{\link{duplicateSample}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
