% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateSample}
\alias{duplicateSample}
\title{De-duplicate by randomly sampling intermediate tibbles}
\usage{
duplicateSample(dupes)
}
\arguments{
\item{dupes}{Intermediate tibble created in \code{\link{duplicateFix}}}
}
\description{
The internal \code{duplicateSample} function is used inside of
\code{\link{duplicateFix}} to deduplicate intermediate tibbles that have been
evaluated using other criteria already.
}
\seealso{
Other deduplication functions: 
\code{\link{duplicateAllOnes}()},
\code{\link{duplicateAllOnesGroupSize}()},
\code{\link{duplicateDecide}()},
\code{\link{duplicateFields}()},
\code{\link{duplicateFinder}()},
\code{\link{duplicateFix}()},
\code{\link{duplicateID}()},
\code{\link{duplicateNewest}()},
\code{\link{duplicatePlot}()},
\code{\link{duplicateRecordType}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{deduplication functions}
