% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicates.R
\name{duplicateSample}
\alias{duplicateSample}
\title{De-duplicate by randomly sampling intermediate tibbles}
\usage{
duplicateSample(dupes)
}
\arguments{
\item{dupes}{Intermediate tibble created in \code{\link{duplicateFix}}}
}
\description{
The internal \code{duplicateSample} function is used inside of \code{\link{duplicateFix}} to deduplicate intermediate tibbles that have been evaluated using other criteria already.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
