% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorLevel_errors_field}
\alias{errorLevel_errors_field}
\title{Calculate error-level errors by field}
\usage{
errorLevel_errors_field(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with \code{\link{proof}} or \code{\link{correct}}}
}
\description{
The internal \code{errorLevel_errors_field} function calculates a summary table of the count of errors and proportion of erroneous values by field.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
