% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorLevelErrorsByState}
\alias{errorLevelErrorsByState}
\title{Calculate error-level errors by state}
\usage{
errorLevelErrorsByState(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}} or \code{\link{correct}}}
}
\description{
The internal \code{errorLevelErrorsByState} function calculates a summary
table of the count of errors, count of correct values, and proportion of
erroneous values by state.
}
\seealso{
Other error-finding functions: 
\code{\link{errorLevelErrorsByField}()},
\code{\link{errorPlotDL}()},
\code{\link{errorPlotFields}()},
\code{\link{errorPlotStates}()},
\code{\link{errorTable}()},
\code{\link{errorTableSummary}()},
\code{\link{pullErrors}()},
\code{\link{redFlags}()}

Other plotting functions: 
\code{\link{duplicatePlot}()},
\code{\link{errorLevelErrorsByField}()},
\code{\link{errorPlotDL}()},
\code{\link{errorPlotFields}()},
\code{\link{errorPlotStates}()},
\code{\link{issuePlot}()},
\code{\link{redFlags}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{error-finding functions}
\concept{plotting functions}
