% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorPlot_states}
\alias{errorPlot_states}
\title{Plot errors by state}
\usage{
errorPlot_states(proofed_data, threshold = NA)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with \code{\link{proof}} or \code{\link{correct}}}

\item{threshold}{Optional. A decimal value above which error proportions should be plotted.}
}
\description{
Create a bar plot of errors by state, either by count or proportion.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
