% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlot_states.R
\name{errorPlot_states}
\alias{errorPlot_states}
\title{Plot errors by state}
\usage{
errorPlot_states(x, threshold)
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{threshold}{Decimal value above which error proportions should be plotted}
}
\description{
Create a bar plot of errors by state, either by count or proportion.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
