% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{errorPlotDL}
\alias{errorPlotDL}
\title{Plot errors across download cycles}
\usage{
errorPlotDL(proofed_data, loc = "all")
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}} or \code{\link{correct}}}

\item{loc}{Which location the error data should be plotted for. Acceptable
values include:
\itemize{
\item "all" - all states
\item a two-letter abbreviation for a US state (excluding HI)
}}
}
\description{
Create a plot of errors per download cycle, either by all states in the data
set or a specific state, province, or territory.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
