% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorTables.R
\name{errorTable}
\alias{errorTable}
\title{Table of errors}
\usage{
errorTable(proofed_data, loc = "all", field = "all")
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}} or \code{\link{correct}}}

\item{loc}{Which location the error data should be tabulated by. Acceptable
values include:
\itemize{
\item a two-letter abbreviation for a US state (excluding HI)
\item "all" - all states
\item "none" - table will not include location in its output
}}

\item{field}{Field the error data should be tabulated by. Acceptable values
include:
\itemize{
\item If loc = "none", field must be "all". Otherwise, choose one of:
\itemize{
\item REF_ALL_FIELDS}
\item "all" - all fields
\item "none" - table will not include field in its output
}}
}
\description{
Create a tibble of existing errors in the data, customizing the output with
value specifications.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
