% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorTable.R
\name{errorTable}
\alias{errorTable}
\title{Table of errors}
\usage{
errorTable(x, loc = "all", field = "all")
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{loc}{Which location the error data should be tabulated by. Acceptable values include:
\itemize{
\item One of the following abbreviations:
\itemize{
\item AL, AK, AZ, AR, CA, CO, CT, DE, DC, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY}
\item all - All states, provinces, and/or territories that exist in the data
\item none - Table will not include location in its output
}}

\item{field}{Field the error data should be tabulated by. Acceptable values include:
\itemize{
\item One of the following fields:
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe_bag, rails_gallinules_bag, cranes, band_tailed_pigeon, brant, seaducks, registration_year, email}
\item all - All fields in the data
\item none - Table will not include field in its output
}}
}
\description{
Create a tibble of existing errors in the data, customizing the output with value specifications.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
