% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorTables.R
\name{errorTableSummary}
\alias{errorTableSummary}
\title{Summary table of errors}
\usage{
errorTableSummary(proofed_data, initial_tbl, loc, field)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with \code{\link{proof}} or \code{\link{correct}}}

\item{initial_tbl}{\code{\link{errorTable}} intermediate object}

\item{loc}{Which location the error data should be tabulated by. Acceptable values include:
\itemize{
\item One of the following abbreviations:
\itemize{
\item AL, AK, AZ, AR, CA, CO, CT, DE, DC, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY}
\item all - All states, provinces, and/or territories that exist in the data
\item none - Table will not include location in its output
}}

\item{field}{Field the error data should be tabulated by. Acceptable values include:
\itemize{
\item If loc = "none", field must be "all". Otherwise, choose one of the following values:
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_yr, email}
\item all - All fields in the data
\item none - Table will not include field in its output
}}
}
\description{
Internal function used inside of \code{\link{errorTable}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
