% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failBTPI}
\alias{failBTPI}
\title{Fail if BTPI permit file contains bad values}
\usage{
failBTPI(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
Internal function that applies to BTPI permit files.
}
\seealso{
Other failure functions: 
\code{\link{failCR}()},
\code{\link{failDLdate}()},
\code{\link{failDLstate}()},
\code{\link{failTF}()},
\code{\link{failYear}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
