% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failCR}
\alias{failCR}
\title{Fail if crane permit file contains bad values}
\usage{
failCR(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
Internal function that applies to crane permit files.
}
\seealso{
Other failure functions: 
\code{\link{failBTPI}()},
\code{\link{failDLdate}()},
\code{\link{failDLstate}()},
\code{\link{failTF}()},
\code{\link{failWidths}()},
\code{\link{failYear}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
