% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failCR}
\alias{failCR}
\title{Fail if crane permit file contains bad values}
\usage{
failCR(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
Internal function that applies to crane permit files.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
