% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failDLdate}
\alias{failDLdate}
\title{Fail if there is an NA in dl_date}
\usage{
failDLdate(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
The internal \code{qDLdate} function is used inside of
\code{\link{write_hip}}.
}
\seealso{
Other failure functions: 
\code{\link{failBTPI}()},
\code{\link{failCR}()},
\code{\link{failDLstate}()},
\code{\link{failProofed}()},
\code{\link{failState}()},
\code{\link{failTF}()},
\code{\link{failWidths}()},
\code{\link{failYear}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
