% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failDLdate}
\alias{failDLdate}
\title{Fail if there is an NA in dl_date}
\usage{
failDLdate(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
The internal \code{qDLdate} function is used inside of
\code{\link{write_hip}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
