% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failDLstate}
\alias{failDLstate}
\title{Fail if there is an NA in dl_state}
\usage{
failDLstate(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
The internal \code{qDLstate} function is used inside of
\code{\link{write_hip}}.
}
\seealso{
Other failure functions: 
\code{\link{failBTPI}()},
\code{\link{failCR}()},
\code{\link{failDLdate}()},
\code{\link{failTF}()},
\code{\link{failYear}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
