% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failProofed}
\alias{failProofed}
\title{Fail if non-proofed data frame supplied}
\usage{
failProofed(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with
\code{\link{proof}}}
}
\description{
Internal function that fails if a tibble without \code{errors} field
(generated by \code{\link{proof}}) is provided.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
