% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failState}
\alias{failState}
\title{Fail if bad state abbreviation supplied}
\usage{
failState(state)
}
\arguments{
\item{state}{Any abbreviation for one of the 49 contiguous states:
\itemize{
\item AL, AK, AZ, AR, CA, CO, CT, DE, FL, GA, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY}}
}
\description{
Internal function that fails if a bad state abbreviation is provided.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
