% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failTF}
\alias{failTF}
\title{Fail if TRUE or FALSE is not supplied}
\usage{
failTF(TF)
}
\arguments{
\item{TF}{Must be TRUE, FALSE, T, or F}
}
\description{
Internal function that fails if something other than TRUE or FALSE is
supplied.
}
\seealso{
Other failure functions: 
\code{\link{failBTPI}()},
\code{\link{failCR}()},
\code{\link{failDLdate}()},
\code{\link{failDLstate}()},
\code{\link{failProofed}()},
\code{\link{failState}()},
\code{\link{failWidths}()},
\code{\link{failYear}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
