% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failWidths}
\alias{failWidths}
\title{Fail if field widths are exceeded by any value}
\usage{
failWidths(corrected_data)
}
\arguments{
\item{corrected_data}{The object created after correcting data with
\code{\link{correct}}}
}
\description{
Internal function that fails inside of \code{\link{write_hip}} if any value
exceeds designated field widths.
}
\seealso{
Other failure functions: 
\code{\link{failBTPI}()},
\code{\link{failCR}()},
\code{\link{failDLdate}()},
\code{\link{failDLstate}()},
\code{\link{failProofed}()},
\code{\link{failState}()},
\code{\link{failTF}()},
\code{\link{failYear}()}

Other writing functions: 
\code{\link{writeReport}()},
\code{\link{write_hip}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
\concept{writing functions}
