% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failYear}
\alias{failYear}
\title{Fail if incorrect year supplied}
\usage{
failYear(year)
}
\arguments{
\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
Internal function that fails if an incorrect year is provided.
}
\seealso{
Other failure functions: 
\code{\link{failBTPI}()},
\code{\link{failCR}()},
\code{\link{failDLdate}()},
\code{\link{failDLstate}()},
\code{\link{failTF}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{failure functions}
