% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fails.R
\name{failyear}
\alias{failyear}
\title{Fail if incorrect year supplied}
\usage{
failyear(year)
}
\arguments{
\item{year}{The year in which the Harvest Information Program data were collected}
}
\description{
Internal function that fails if an incorrect year is provided.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
