% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileCheck.R
\name{fileCheck}
\alias{fileCheck}
\title{Check for repeat files}
\usage{
fileCheck(raw_path, processed_path)
}
\arguments{
\item{raw_path}{Directory of the download folder containing HIP .txt files}

\item{processed_path}{Directory of the folder containing processed files}
}
\description{
Check if any files in the input folder have already been written to processed
folder.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
