% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDuplicates.R
\name{fixDuplicates}
\alias{fixDuplicates}
\title{Fix duplicates}
\usage{
fixDuplicates(x)
}
\arguments{
\item{x}{A current data table created by \code{\link{issueCheck}}}
}
\description{
Resolve duplicate HIP records.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
