% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fixDuplicates.R
\name{fixDuplicates}
\alias{fixDuplicates}
\title{Fix duplicates}
\usage{
fixDuplicates(x)
}
\arguments{
\item{x}{A cleaned data table created by \code{\link{clean}}}
}
\description{
Consolidate duplicate records that stem from HIP data and special permit information into one row. If other duplicates exist not due to this specific reason, delete the records from the data.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
