% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{formatZip}
\alias{formatZip}
\title{Format zip codes}
\usage{
formatZip(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{formatZip} function formats zip codes.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
