% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{futureDateMessage}
\alias{futureDateMessage}
\title{Return message if future issue_date values are detected}
\usage{
futureDateMessage(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{futureDateMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if future \code{issue_date} values are detected.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
