% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadAddress}
\alias{getBadAddress}
\title{Get bad address values}
\usage{
getBadAddress(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadAddress} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{address} values that are not expected. Used by \code{\link{qAddress}}.
}
\seealso{
Other proofing functions: 
\code{\link{getBadBirthDate}()},
\code{\link{getBadCity}()},
\code{\link{getBadFirstName}()},
\code{\link{getBadHuntMigBirds}()},
\code{\link{getBadLastName}()},
\code{\link{getBadMiddle}()},
\code{\link{getBadRegYear}()},
\code{\link{getBadState}()},
\code{\link{getBadSuffix}()},
\code{\link{getBadTitle}()},
\code{\link{getBadZIP}()},
\code{\link{proof}()},
\code{\link{proofBadEmails}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
