% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadAddress}
\alias{getBadAddress}
\title{Get bad address values}
\usage{
getBadAddress(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadAddress} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{address} values that are not expected. Used by \code{\link{qAddress}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
