% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadBirthDate}
\alias{getBadBirthDate}
\title{Get bad birth date values}
\usage{
getBadBirthDate(data, year)
}
\arguments{
\item{data}{Harvest Information Program registration data}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The internal \code{getBadBirthDate} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{birth_date} values that are not expected. We expect birth year to be
between 0 and 100 years ago. Used by \code{\link{qBirthDate}}.
}
\seealso{
Other proofing functions: 
\code{\link{getBadAddress}()},
\code{\link{getBadCity}()},
\code{\link{getBadFirstName}()},
\code{\link{getBadHuntMigBirds}()},
\code{\link{getBadLastName}()},
\code{\link{getBadMiddle}()},
\code{\link{getBadRegYear}()},
\code{\link{getBadState}()},
\code{\link{getBadSuffix}()},
\code{\link{getBadTitle}()},
\code{\link{getBadZIP}()},
\code{\link{proof}()},
\code{\link{proofBadEmails}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
