% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadCity}
\alias{getBadCity}
\title{Get bad city values}
\usage{
getBadCity(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadCity} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{city} values that are not expected. Used by \code{\link{qCity}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
