% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadFirstName}
\alias{getBadFirstName}
\title{Get bad first name values}
\usage{
getBadFirstName(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadFirstName} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{firstname} values that are not expected. Used by
\code{\link{qFirstName}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
