% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadHuntMigBirds}
\alias{getBadHuntMigBirds}
\title{Get bad hunt mig birds values}
\usage{
getBadHuntMigBirds(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadHuntMigBirds} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{hunt_mig_birds} values that are not expected.
}
\seealso{
Other proofing functions: 
\code{\link{getBadAddress}()},
\code{\link{getBadBirthDate}()},
\code{\link{getBadCity}()},
\code{\link{getBadFirstName}()},
\code{\link{getBadLastName}()},
\code{\link{getBadMiddle}()},
\code{\link{getBadRegYear}()},
\code{\link{getBadState}()},
\code{\link{getBadSuffix}()},
\code{\link{getBadTitle}()},
\code{\link{getBadZIP}()},
\code{\link{proof}()},
\code{\link{proofBadEmails}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
