% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadHuntMigBirds}
\alias{getBadHuntMigBirds}
\title{Get bad hunt mig birds values}
\usage{
getBadHuntMigBirds(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadHuntMigBirds} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{hunt_mig_birds} values that are not expected.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
