% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadLastName}
\alias{getBadLastName}
\title{Get bad last name values}
\usage{
getBadLastName(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadLastName} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{lastname} values that are not expected. Used by
\code{\link{qLastName}}.
}
\seealso{
Other proofing functions: 
\code{\link{getBadAddress}()},
\code{\link{getBadBirthDate}()},
\code{\link{getBadCity}()},
\code{\link{getBadFirstName}()},
\code{\link{getBadHuntMigBirds}()},
\code{\link{getBadMiddle}()},
\code{\link{getBadRegYear}()},
\code{\link{getBadState}()},
\code{\link{getBadSuffix}()},
\code{\link{getBadTitle}()},
\code{\link{getBadZIP}()},
\code{\link{proof}()},
\code{\link{proofBadEmails}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
