% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadLastName}
\alias{getBadLastName}
\title{Get bad last name values}
\usage{
getBadLastName(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadLastName} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{lastname} values that are not expected. Used by
\code{\link{qLastName}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
