% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadRegYear}
\alias{getBadRegYear}
\title{Get bad registration year values}
\usage{
getBadRegYear(data, year)
}
\arguments{
\item{data}{Harvest Information Program registration data}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The internal \code{getBadRegYear} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{registration_yr} values that are not expected. We expect the
registration_yr to equal the current hunting season.
}
\seealso{
Other proofing functions:
\code{\link[=getBadAddress]{getBadAddress()}},
\code{\link[=getBadBirthDate]{getBadBirthDate()}},
\code{\link[=getBadCity]{getBadCity()}},
\code{\link[=getBadFirstName]{getBadFirstName()}},
\code{\link[=getBadHuntMigBirds]{getBadHuntMigBirds()}},
\code{\link[=getBadLastName]{getBadLastName()}},
\code{\link[=getBadMiddle]{getBadMiddle()}},
\code{\link[=getBadState]{getBadState()}},
\code{\link[=getBadSuffix]{getBadSuffix()}},
\code{\link[=getBadTitle]{getBadTitle()}},
\code{\link[=getBadZIP]{getBadZIP()}},
\code{\link[=proof]{proof()}},
\code{\link[=proofBadEmails]{proofBadEmails()}}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
