% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadRegYear}
\alias{getBadRegYear}
\title{Get bad registration year values}
\usage{
getBadRegYear(data, year)
}
\arguments{
\item{data}{Harvest Information Program registration data}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The internal \code{getBadRegistrationYear} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{registration_yr} values that are not expected. We expect the
registration_yr to equal the current hunting season.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
