% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadState}
\alias{getBadState}
\title{Get bad state values}
\usage{
getBadState(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadState} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{state} values that are not expected. Used by \code{\link{qState}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
