% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadSuffix}
\alias{getBadSuffix}
\title{Get bad suffix values}
\usage{
getBadSuffix(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadSuffix} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{suffix} values that are not expected. Used by \code{\link{qSuffix}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
