% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadTitle}
\alias{getBadTitle}
\title{Get bad title values}
\usage{
getBadTitle(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadTitle} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{title} values that are not expected. Used by \code{\link{qTitle}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
