% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{getBadZIP}
\alias{getBadZIP}
\title{Get bad zip code values}
\usage{
getBadZIP(data)
}
\arguments{
\item{data}{Harvest Information Program registration data}
}
\description{
The internal \code{getBadZIP} function is used inside of
\code{\link{proof}} and \code{\link{qualityMessages}} to filter to
\code{zip} values that are not expected. Used by \code{\link{qZIP}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
