% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyphs.R
\name{glyphFinder}
\alias{glyphFinder}
\title{Find non-UTF-8 glyphs/characters in a field}
\usage{
glyphFinder(pulled_data, field)
}
\arguments{
\item{pulled_data}{The tibble created after reading in data with \code{\link{read_hip}}}

\item{field}{Field that should be checked for non-UTF-8 characters. One of the fields from the following list may be supplied:
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_yr, email}}
}
\description{
The internal \code{glyphFinder} function pulls non-UTF-8 characters in a field.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
