% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyphFinder.R
\name{glyphFinder}
\alias{glyphFinder}
\title{Find non-UTF-8 glyphs/character}
\usage{
glyphFinder(data, field)
}
\arguments{
\item{data}{HIP data table at any point during pre-processing}

\item{field}{Field that should be checked for non-UTF-8 characters. One of the fields from the following list may be supplied:
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe, rails_gallinules, cranes, band_tailed_pigeon, brant, seaducks, registration_year, email}}
}
\description{
Pull and view non-UTF-8 characters in a field when R throws error 'invalid UTF-8 byte sequence detected'.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
