% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{idBlankFiles}
\alias{idBlankFiles}
\title{Identify blank files}
\usage{
idBlankFiles(filelist)
}
\arguments{
\item{filelist}{The file list tibble created by \code{\link{listFiles}}}
}
\description{
The internal \code{idBlankFiles} function is used inside of
\code{\link{read_hip}} to identify files in the list that contain no data.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
