% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{inLinePermitDNHMessage}
\alias{inLinePermitDNHMessage}
\title{In-line permit did-not-hunt message}
\usage{
inLinePermitDNHMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{inLinePermitDNHMessage} function returns a message for in-line permit records from OR or WA that indicate they did not hunt.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
