% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{interStateDuplicatesMessage}
\alias{interStateDuplicatesMessage}
\title{Inter-state duplicates message}
\usage{
interStateDuplicatesMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{interStateDuplicatesMessage} function returns a message
for duplicates found among files submitted from different states.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
