% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{invalidDateMessage}
\alias{invalidDateMessage}
\title{Return message if invalid issue_date values are detected}
\usage{
invalidDateMessage(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{invalidDateMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if invalid \code{issue_date} values are detected.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
