% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/investigate.R
\name{investigate}
\alias{investigate}
\title{Investigate data}
\usage{
investigate(x, dl_state, dl_date, species)
}
\arguments{
\item{x}{The object created after tidying data with \code{\link{tidy}}}

\item{dl_state}{The download state in question}

\item{dl_date}{The download date in question}

\item{species}{The bird group in question}
}
\description{
After checking the data with \code{\link{validate}}, look into reported repeated values to determine if they are worrisome.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
