% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issueCheck}
\alias{issueCheck}
\title{Check issue dates and license years}
\usage{
issueCheck(clean_data, year, plot = FALSE)
}
\arguments{
\item{clean_data}{The object created after cleaning data with
\code{\link{clean}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}

\item{plot}{Create a plot? Default is FALSE}
}
\description{
After cleaning the data with \code{\link{clean}}, ensure each record is
assigned the appropriate \code{registration_yr}.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueDecide}()},
\code{\link{issueMessages}()},
\code{\link{issuePlot}()},
\code{\link{issuePlotDateLabel}()},
\code{\link{issuePrint}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{timeTravelMessage}()},
\code{\link{twoSeasonMessage}()},
\code{\link{zeroDateMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
