% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issueDecide}
\alias{issueDecide}
\title{Assign decisions to records using issue date and registration year}
\usage{
issueDecide(clean_data, year)
}
\arguments{
\item{clean_data}{The object created after cleaning data with
\code{\link{clean}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
The internal \code{issueDecide} function is used inside of
\code{\link{issueAssign}} and \code{\link{issueCheck}} to create the
\code{decision} field, which assigns records as current, past, future, etc;
no data changes are made other than creating the \code{deicison} field.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueCheck}()},
\code{\link{issueMessages}()},
\code{\link{issuePlot}()},
\code{\link{issuePrint}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{timeTravelMessage}()},
\code{\link{twoSeasonMessage}()},
\code{\link{zeroDateMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
