% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issueMessages}
\alias{issueMessages}
\title{Return messages to console for issueCheck insights}
\usage{
issueMessages(clean_data, issue_assignments)
}
\arguments{
\item{clean_data}{The product of \code{\link{clean}}}

\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{issueMessages} function is used inside of
\code{\link{issueCheck}} to return messages for bad issue_date values, count
of past registrations, count of future registrations, and 2-season overlap
registrations.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
