% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePlot}
\alias{issuePlot}
\title{Plot issue date errors}
\usage{
issuePlot(issue_assignments, year)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
The internal \code{issuePlot} function plots bad \code{issue_date} values.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueCheck}()},
\code{\link{issueDecide}()},
\code{\link{issueMessages}()},
\code{\link{issuePrint}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{timeTravelMessage}()},
\code{\link{twoSeasonMessage}()},
\code{\link{zeroDateMessage}()}

Other plotting functions: 
\code{\link{duplicatePlot}()},
\code{\link{errorLevelErrorsByField}()},
\code{\link{errorLevelErrorsByState}()},
\code{\link{errorPlotDL}()},
\code{\link{errorPlotFields}()},
\code{\link{errorPlotStates}()},
\code{\link{redFlags}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
\concept{plotting functions}
