% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePlot}
\alias{issuePlot}
\title{Plot issue date errors}
\usage{
issuePlot(x, year)
}
\arguments{
\item{x}{The object created by \code{\link{issueAssign}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
The internal \code{issuePlot} function plots the output of \code{\link{issueAssign}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
