% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePlot}
\alias{issuePlot}
\title{Plot issue date errors}
\usage{
issuePlot(issue_assignments, year)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
The internal \code{issuePlot} function plots bad \code{issue_date} values.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
