% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePlotDateLabel}
\alias{issuePlotDateLabel}
\title{Label dates for issue date plot}
\usage{
issuePlotDateLabel(x)
}
\arguments{
\item{x}{Vector of dates}
}
\description{
The internal \code{issuePlotDateLabel} function labels x-axis dates for
\code{\link{issuePlot}}.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueCheck}()},
\code{\link{issueDecide}()},
\code{\link{issueMessages}()},
\code{\link{issuePlot}()},
\code{\link{issuePrint}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{timeTravelMessage}()},
\code{\link{twoSeasonMessage}()},
\code{\link{zeroDateMessage}()}

Other plotting functions: 
\code{\link{duplicatePlot}()},
\code{\link{errorLevelErrorsByField}()},
\code{\link{errorLevelErrorsByState}()},
\code{\link{errorPlotDL}()},
\code{\link{errorPlotFields}()},
\code{\link{errorPlotStates}()},
\code{\link{issuePlot}()},
\code{\link{redFlags}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
\concept{plotting functions}
