% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePrint}
\alias{issuePrint}
\title{Print results of issueCheck changes}
\usage{
issuePrint(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{issuePrint} function is used inside of
\code{\link{issueCheck}} to print the number of registration types (past,
future, current, etc) per download state and registration year.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueCheck}()},
\code{\link{issueDecide}()},
\code{\link{issueMessages}()},
\code{\link{issuePlot}()},
\code{\link{issuePlotDateLabel}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{timeTravelMessage}()},
\code{\link{twoSeasonMessage}()},
\code{\link{zeroDateMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
