% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{issuePrint}
\alias{issuePrint}
\title{Print results of issueCheck changes}
\usage{
issuePrint(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{issuePrint} function is used inside of
\code{\link{issueCheck}} to print the number of registration types (past,
future, current, etc) per download state and registration year.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
