% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/migbirdHIP-package.R
\docType{package}
\name{migbirdHIP-package}
\alias{migbirdHIP}
\alias{migbirdHIP-package}
\title{migbirdHIP: Process USFWS Migratory Bird Harvest Information Program Data in R}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Functions designed to process and visualize data submitted to the USFWS Harvest Information Program.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.fws.gov/project/migratory-bird-hunter-survey-diary-survey}
}

}
\author{
\strong{Maintainer}: Abby Walter \email{abby_walter@fws.gov} (\href{https://orcid.org/0000-0003-4017-5596}{ORCID})

}
\keyword{internal}
