% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{missingEmailsMessage}
\alias{missingEmailsMessage}
\title{Return message if all emails are missing or are exactly the same in a file
with more than 10 registrations}
\usage{
missingEmailsMessage(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{missingEmailsMessage} function is used inside of
\code{\link{qualityMessages}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
