% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{missingPIIFilter}
\alias{missingPIIFilter}
\title{Missing PII filter}
\usage{
missingPIIFilter(raw_df)
}
\arguments{
\item{raw_df}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
The internal \code{missingPIIFilter} function filters out HIP registrations that are missing critical pieces of contact information.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
