% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{missingPIIMessage}
\alias{missingPIIMessage}
\title{Return message for records with blank or NA values in firstname, lastname,
state, or birth date}
\usage{
missingPIIMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{missingPIIMessage} function is used inside of
\code{\link{readMessages}}
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
