% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{moveSuffixes}
\alias{moveSuffixes}
\title{Move suffixes}
\usage{
moveSuffixes(raw_df)
}
\arguments{
\item{raw_df}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
The internal \code{moveSuffixes} function moves suffixes from first name or last name columns into the suffix column and performs other cleaning steps. This function catches values from 1 to 20 in Roman numerals and numeric, excluding
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
