% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{naBagsMessage}
\alias{naBagsMessage}
\title{Return message if any record has an NA in every bag field}
\usage{
naBagsMessage(pulled_data)
}
\arguments{
\item{pulled_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{naBagsMessage} function is used inside of \code{\link{readMessages}}
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
