% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nDropped.R
\name{nDroppedCurrent}
\alias{nDroppedCurrent}
\title{Summary table of how many registrations were dropped by issueCheck}
\usage{
nDroppedCurrent(clean_data, current_data, year)
}
\arguments{
\item{clean_data}{The object created after cleaning data with
\code{\link{clean}}}

\item{current_data}{The object created after filtering to current data with
\code{\link{issueCheck}}}

\item{year}{The year of the HIP season (e.g. 2022 for the 2022-2023 season)}
}
\description{
Internal function that helps \code{\link{nDropped}} by summarizing the number
of dropped registrations by \code{\link{issueCheck}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
