% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{nonDigitBagsMessage}
\alias{nonDigitBagsMessage}
\title{Return message if any record contains a bag value that is not a 1-digit number}
\usage{
nonDigitBagsMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{nonDigitBagsMessage} function is used inside of \code{\link{readMessages}}
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
