% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{nonResidentMessage}
\alias{nonResidentMessage}
\title{High non-resident hunter proportion message}
\usage{
nonResidentMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{nonResidentMessage} function returns a message for files
with 10\% or more of \code{state} values that do not match \code{dl_state}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
