% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographics.R
\name{outOfStateHunters}
\alias{outOfStateHunters}
\title{Find out-of-state hunters}
\usage{
outOfStateHunters(proofed_data, return = "plot")
}
\arguments{
\item{proofed_data}{The object created after error flagging data with \code{\link{proof}} or \code{\link{correct}}}

\item{return}{Return a "plot" (default) or "table"}
}
\description{
Create a tibble or plot of hunters who have registered outside of their home state.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
