% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfStateHunters.R
\name{outOfStateHunters}
\alias{outOfStateHunters}
\title{Find out-of-state hunters}
\usage{
outOfStateHunters(x)
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}
}
\description{
Create a tibble and plot of hunters who have reported hunting migratory waterfowl outside of the state associated with their address.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
