% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outOfStateHunters.R
\name{outOfStateHunters}
\alias{outOfStateHunters}
\title{Find out-of-state hunters}
\usage{
outOfStateHunters(x, output = "plot")
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{output}{"plot" (defualt) or "table"}
}
\description{
Create a tibble or plot of hunters who have registered outside of their home state.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
