% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{permitFileBagsMessage}
\alias{permitFileBagsMessage}
\title{Permit file non-zero bag values message}
\usage{
permitFileBagsMessage(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{permitFileBagsMessage} function returns a message for
records from permit file state/species combinations with non-zero bag values.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
