% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proof}
\alias{proof}
\title{Proof data and flag errors}
\usage{
proof(deduplicated_data, year)
}
\arguments{
\item{deduplicated_data}{The object created after deduplicating data with
\code{\link{duplicateFix}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
After cleaning the data with \code{\link{clean}}, compare each field to an
expected range of values and flag non-conforming values in a new "errors"
column.
}
\seealso{
Other proofing functions: 
\code{\link{getBadAddress}()},
\code{\link{getBadBirthDate}()},
\code{\link{getBadCity}()},
\code{\link{getBadFirstName}()},
\code{\link{getBadHuntMigBirds}()},
\code{\link{getBadLastName}()},
\code{\link{getBadMiddle}()},
\code{\link{getBadRegYear}()},
\code{\link{getBadState}()},
\code{\link{getBadSuffix}()},
\code{\link{getBadTitle}()},
\code{\link{getBadZIP}()},
\code{\link{proofBadEmails}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
