% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proof}
\alias{proof}
\title{Flag errors}
\usage{
proof(x, year)
}
\arguments{
\item{x}{The object created after cleaning data with \code{\link{clean}}}

\item{year}{The year in which the Harvest Information Program data were collected}
}
\description{
After cleaning the data with \code{\link{clean}}, compare each field to an expected range of values and flag non-conforming values in a new "errors" column.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
