% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proofBadEmails}
\alias{proofBadEmails}
\title{Proof bad emails}
\usage{
proofBadEmails(keyed_data)
}
\arguments{
\item{keyed_data}{An tibble used internally in \code{\link{proof}}}
}
\description{
The internal \code{proofBadEmails} function is used inside of
\code{\link{proof}} to find poorly formatted or intentionally obfuscative
email addresses. Email addresses that match \code{REGEX_EMAIL} or are
\code{NA} are considered good emails.
}
\seealso{
Other proofing functions: 
\code{\link{getBadAddress}()},
\code{\link{getBadBirthDate}()},
\code{\link{getBadCity}()},
\code{\link{getBadFirstName}()},
\code{\link{getBadHuntMigBirds}()},
\code{\link{getBadLastName}()},
\code{\link{getBadMiddle}()},
\code{\link{getBadRegYear}()},
\code{\link{getBadState}()},
\code{\link{getBadSuffix}()},
\code{\link{getBadTitle}()},
\code{\link{getBadZIP}()},
\code{\link{proof}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{proofing functions}
