% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proof.R
\name{proofBadEmails}
\alias{proofBadEmails}
\title{Proof bad emails}
\usage{
proofBadEmails(keyed_data)
}
\arguments{
\item{keyed_data}{An tibble used internally in \code{\link{proof}}}
}
\description{
The internal \code{proofBadEmails} function is used inside of
\code{\link{proof}} to find poorly formatted or intentionally obfuscative
email addresses. Email addresses that match \code{REGEX_EMAIL} or are
\code{NA} are considered good emails.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
