% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pullErrors.R
\name{pullErrors}
\alias{pullErrors}
\title{Pull flagged errors}
\usage{
pullErrors(x, error, distinct = TRUE)
}
\arguments{
\item{x}{A proofed data table created by \code{\link{proof}}}

\item{error}{Field that should be pulled. One of the fields from the following list may be supplied:
\itemize{
\item title, firstname, middle, lastname, suffix, address, city, state, zip, birth_date, issue_date, hunt_mig_birds, ducks_bag, geese_bag, dove_bag, woodcock_bag, coots_snipe_bag, rails_gallinules_bag, cranes, band_tailed_pigeon, brant, seaducks, registration_year, email}}

\item{distinct}{If FALSE, returns all error vvalues ; if TRUE (default), only returns unique values.}
}
\description{
Pull and view errors that have been flagged for a particular field. This function allows you to easily see what the \code{\link{proof}} function has determined to be unacceptable data.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
