% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qAddress}
\alias{qAddress}
\title{Bad address message}
\usage{
qAddress(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qAddress} function returns a message for bad
\code{address} values. Uses internal helper \code{\link{getBadAddress}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
