% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qBags}
\alias{qBags}
\title{Bad bag messages}
\usage{
qBags(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qBags} function returns a messages for bad bag values,
using internal helper functions \code{\link{zeroBagsMessage}},
\code{\link{naBagsMessage}}, and \code{\link{nonDigitBagsMessage}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
