% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qBirthDate}
\alias{qBirthDate}
\title{Bad birth dates message}
\usage{
qBirthDate(raw_data, year)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The internal \code{qBirthDate} function returns a message for
\code{birth_date} values that are incorrectly formatted or improbable. Uses
internal helper \code{\link{getBadBirthDate}}.
}
\seealso{
Other quality functions: 
\code{\link{inLinePermitDNHMessage}()},
\code{\link{interStateDuplicatesMessage}()},
\code{\link{missingEmailsMessage}()},
\code{\link{missingPIIMessage}()},
\code{\link{naBagsMessage}()},
\code{\link{nonDigitBagsMessage}()},
\code{\link{nonResidentMessage}()},
\code{\link{permitFileBagsMessage}()},
\code{\link{qAddress}()},
\code{\link{qBags}()},
\code{\link{qBirthDateRange}()},
\code{\link{qCity}()},
\code{\link{qFirstName}()},
\code{\link{qHuntMigBirds}()},
\code{\link{qIssueDateRange}()},
\code{\link{qLastName}()},
\code{\link{qMiddle}()},
\code{\link{qRegistrationYear}()},
\code{\link{qState}()},
\code{\link{qSuffix}()},
\code{\link{qSummary}()},
\code{\link{qTitle}()},
\code{\link{qZIP}()},
\code{\link{qualityMessages}()},
\code{\link{testRecordMessage}()},
\code{\link{zeroBagsMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
