% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qBirthDate}
\alias{qBirthDate}
\title{Bad birth dates message}
\usage{
qBirthDate(raw_data, year)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The internal \code{qBirthDate} function returns a message for
\code{birth_date} values that are incorrectly formatted or improbable. Uses
internal helper \code{\link{getBadBirthDate}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
