% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qBirthDateRange}
\alias{qBirthDateRange}
\title{Birth date range message}
\usage{
qBirthDateRange(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{qBirthDateRange} function returns a message for
\code{birth_date} values that do not span a period of time more than 365
days.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
