% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qCity}
\alias{qCity}
\title{Bad city message}
\usage{
qCity(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qCity} function returns a message for bad \code{city}
values. Uses internal helper \code{\link{getBadCity}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
