% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qFirstName}
\alias{qFirstName}
\title{Bad first name message}
\usage{
qFirstName(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qFirstName} function returns a message for bad
\code{firstname} values. Uses internal helper \code{\link{getBadFirstName}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
