% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qHuntMigBirds}
\alias{qHuntMigBirds}
\title{Return message if records with bad values for hunt_mig_birds are detected}
\usage{
qHuntMigBirds(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qHuntMigBirds} function is used inside of
\code{\link{qualityMessages}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
