% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qLastName}
\alias{qLastName}
\title{Bad last name message}
\usage{
qLastName(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qLastName} function returns a message for bad
\code{lastname} values. Uses internal helper \code{\link{getBadLastName}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
