% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qMiddle}
\alias{qMiddle}
\title{Bad middle initial message}
\usage{
qMiddle(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qMiddle} function returns a message for bad \code{middle}
values. Uses internal helper \code{\link{getBadMiddle}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
