% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qRegistrationYear}
\alias{qRegistrationYear}
\title{Bad registration_yr message}
\usage{
qRegistrationYear(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qRegistrationYear} function returns a message for
\code{registration_yr} values that are not equal to REF_CURRENT_SEASON or
REF_CURRENT_SEASON + 1.
}
\seealso{
Other quality functions: 
\code{\link{inLinePermitDNHMessage}()},
\code{\link{interStateDuplicatesMessage}()},
\code{\link{missingEmailsMessage}()},
\code{\link{missingPIIMessage}()},
\code{\link{naBagsMessage}()},
\code{\link{nonDigitBagsMessage}()},
\code{\link{nonResidentMessage}()},
\code{\link{permitFileBagsMessage}()},
\code{\link{qAddress}()},
\code{\link{qBags}()},
\code{\link{qBirthDate}()},
\code{\link{qBirthDateRange}()},
\code{\link{qCity}()},
\code{\link{qFirstName}()},
\code{\link{qHuntMigBirds}()},
\code{\link{qIssueDateRange}()},
\code{\link{qLastName}()},
\code{\link{qMiddle}()},
\code{\link{qState}()},
\code{\link{qSuffix}()},
\code{\link{qSummary}()},
\code{\link{qTitle}()},
\code{\link{qZIP}()},
\code{\link{qualityMessages}()},
\code{\link{testRecordMessage}()},
\code{\link{zeroBagsMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
