% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qState}
\alias{qState}
\title{Bad state message}
\usage{
qState(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qState} function returns a message for bad \code{state}
values. Uses internal helper \code{\link{getBadState}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
