% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qSuffix}
\alias{qSuffix}
\title{Bad suffix message}
\usage{
qSuffix(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qSuffix} function returns a message for bad \code{suffix}
values. Uses internal helper \code{\link{getBadSuffix}}.
}
\seealso{
Other quality functions: 
\code{\link{inLinePermitDNHMessage}()},
\code{\link{interStateDuplicatesMessage}()},
\code{\link{missingEmailsMessage}()},
\code{\link{missingPIIMessage}()},
\code{\link{naBagsMessage}()},
\code{\link{nonDigitBagsMessage}()},
\code{\link{nonResidentMessage}()},
\code{\link{permitFileBagsMessage}()},
\code{\link{qAddress}()},
\code{\link{qBags}()},
\code{\link{qBirthDate}()},
\code{\link{qBirthDateRange}()},
\code{\link{qCity}()},
\code{\link{qFirstName}()},
\code{\link{qHuntMigBirds}()},
\code{\link{qIssueDateRange}()},
\code{\link{qLastName}()},
\code{\link{qMiddle}()},
\code{\link{qRegistrationYear}()},
\code{\link{qState}()},
\code{\link{qSummary}()},
\code{\link{qTitle}()},
\code{\link{qZIP}()},
\code{\link{qualityMessages}()},
\code{\link{testRecordMessage}()},
\code{\link{zeroBagsMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
