% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qSuffix}
\alias{qSuffix}
\title{Bad suffix message}
\usage{
qSuffix(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qSuffix} function returns a message for bad \code{suffix}
values. Uses internal helper \code{\link{getBadSuffix}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
