% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qSummary}
\alias{qSummary}
\title{Quality summary}
\usage{
qSummary(bad_data)
}
\arguments{
\item{bad_data}{Must contain file_size field}
}
\description{
The internal \code{qSummary} function ...
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
