% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qTitle}
\alias{qTitle}
\title{Bad title message}
\usage{
qTitle(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qTitle} function returns a message for \code{title} values
that are not assigned to the correct registrations. Uses internal helper
\code{\link{getBadTitle}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
