% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qualityMessages}
\alias{qualityMessages}
\title{Return messages to console for common or catastrophic data quality issues}
\usage{
qualityMessages(raw_data, year)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}

\item{year}{The year in which the Harvest Information Program data were
collected}
}
\description{
The \code{qualityMessages} function is used to return messages for
missing PII, missing email addresses, all-zero bag records, non-numeric bag
values, NAs in dl_state, NAs in dl_date, bad title assignments, high
proportions of non-resident hunters in a file, inter-state duplicates, low
range of issue dates in a file, and low range of birth dates in a file.
}
\seealso{
Other quality functions: 
\code{\link{inLinePermitDNHMessage}()},
\code{\link{interStateDuplicatesMessage}()},
\code{\link{missingEmailsMessage}()},
\code{\link{missingPIIMessage}()},
\code{\link{naBagsMessage}()},
\code{\link{nonDigitBagsMessage}()},
\code{\link{nonResidentMessage}()},
\code{\link{permitFileBagsMessage}()},
\code{\link{qAddress}()},
\code{\link{qBags}()},
\code{\link{qBirthDate}()},
\code{\link{qBirthDateRange}()},
\code{\link{qCity}()},
\code{\link{qFirstName}()},
\code{\link{qHuntMigBirds}()},
\code{\link{qIssueDateRange}()},
\code{\link{qLastName}()},
\code{\link{qMiddle}()},
\code{\link{qRegistrationYear}()},
\code{\link{qState}()},
\code{\link{qSuffix}()},
\code{\link{qSummary}()},
\code{\link{qTitle}()},
\code{\link{qZIP}()},
\code{\link{testRecordMessage}()},
\code{\link{zeroBagsMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{quality functions}
