% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualityCheck.R
\name{qualityMessages}
\alias{qualityMessages}
\title{Return messages to console for common or catastrophic data quality issues}
\usage{
qualityMessages(raw_data)
}
\arguments{
\item{raw_data}{The product of \code{\link{read_hip}}}
}
\description{
The internal \code{qualityMessages} function is used inside of
\code{\link{qualityCheck}} to return messages for missing PII, missing email
addresses, all-zero bag records, non-numeric bag values, NAs in dl_state, and
NAs in dl_date.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
