% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{qZIP}
\alias{qZIP}
\title{Bad zip code message}
\usage{
qZIP(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{qZIP} function returns a message for bad \code{zip}
values. Uses internal helper \code{\link{getBadZIP}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
