% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{read_hip}
\alias{read_hip}
\title{Read in data}
\usage{
read_hip(path)
}
\arguments{
\item{path}{File path to the folder containing HIP .txt files}
}
\description{
Compile data from state-exported text files by providing a path to the download directory.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
