% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{read_hip}
\alias{read_hip}
\title{Read in data}
\usage{
read_hip(path, state = NA, season = FALSE)
}
\arguments{
\item{path}{File path to the folder containing HIP .txt files}

\item{state}{When specified, reads in download data from a specified state. Must match one of the following two-letter abbreviations:
\itemize{
\item AL, AK, AZ, AR, CA, CO, CT, DE, DC, FL, GA, HI, ID, IL, IN, IA, KS, KY, LA, ME, MD, MA, MI, MN, MS, MO, MT, NE, NV, NH, NJ, NM, NY, NC, ND, OH, OK, OR, PA, RI, SC, SD, TN, TX, UT, VT, VA, WA, WV, WI, WY, AS, GU, MP, PR, VI, UM, FM, MH, PW, AA, AE, AP, CM, CZ, NB, PI, TT, ON, QC, NS, NB, MB, BC, PE, SK, AB, NL}}

\item{season}{If set as TRUE, selects only folders starting with "DL" in a a season's upper-level directory}
}
\description{
Compile data from state-exported text files by providing a path to the download directory.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
