% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_hip.R
\name{readTimeMessage}
\alias{readTimeMessage}
\title{Read time message}
\usage{
readTimeMessage(raw_data, readtime)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}

\item{readtime}{The difference between end time and start time}
}
\description{
The internal \code{readTimeMessage} function returns a message to the console
informing the user about how long it took to read in the data with
\code{\link{read_hip}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
