% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/errorPlots.R
\name{recordLevel_errors_state}
\alias{recordLevel_errors_state}
\title{Calculate record-level errors by state}
\usage{
recordLevel_errors_state(proofed_data)
}
\arguments{
\item{proofed_data}{The object created after error flagging data with \code{\link{proof}} or \code{\link{correct}}}
}
\description{
The internal \code{recordLevel_errors_state} function calculates a summary table of the count of records with errors, count of records with no errors, and proportion of erroneous records by state.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
