% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{regYearEditMessage}
\alias{regYearEditMessage}
\title{Return message for changed registration_yr values}
\usage{
regYearEditMessage(clean_data, issue_assignments)
}
\arguments{
\item{clean_data}{The product of \code{\link{clean}}}

\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{regYearEditMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
for how many \code{registration_yr} values were changed.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
