% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFiles.R
\name{renameFiles}
\alias{renameFiles}
\title{Raw HIP file name correction}
\usage{
renameFiles(path)
}
\arguments{
\item{path}{Directory to download folder containing new HIP files}
}
\description{
This function overwrites HIP filenames. Files in the supplied directory are renamed by converting the Julian date to YYYYMMDD format. State abbreviations that are in lowercase format are capitalized.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
