% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{shiftCheck}
\alias{shiftCheck}
\title{Check for frame shifts}
\usage{
shiftCheck(data)
}
\arguments{
\item{data}{A raw data table created by \code{\link{read_hip}}}
}
\description{
Find and print any rows that have a line shift error with number of positions shifted.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
