% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shifts.R
\name{shiftFinder}
\alias{shiftFinder}
\title{Find frame shifts}
\usage{
shiftFinder(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{shiftFinder} function finds any rows that have a line
shift error.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
