% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strataCheck.R
\name{strataCheck}
\alias{strataCheck}
\title{Check state strata}
\usage{
strataCheck(x)
}
\arguments{
\item{x}{The object created after cleaning data with \code{\link{clean}}}
}
\description{
After cleaning the data with \code{\link{clean}}, ensure there are no new strata that have been introduced by a state to the species groups.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
