% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sumLines.R
\name{sumLines}
\alias{sumLines}
\title{Sum lines of new data}
\usage{
sumLines(path)
}
\arguments{
\item{path}{File path to the download folder containing HIP .txt files}
}
\description{
Get the sum of the number of lines in new download files.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
