% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quality.R
\name{testRecordMessage}
\alias{testRecordMessage}
\title{Return message if test record is found}
\usage{
testRecordMessage(raw_data)
}
\arguments{
\item{raw_data}{The object created after reading in data with
\code{\link{read_hip}}}
}
\description{
The internal \code{testRecordMessage} function is used inside of
\code{\link{qualityMessages}}.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
