% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{timeTravelMessage}
\alias{timeTravelMessage}
\title{Return message if an issue date is after the file was submitted}
\usage{
timeTravelMessage(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{timeTravelMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if \code{issue_date} values are after the date the file was submitted.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueCheck}()},
\code{\link{issueDecide}()},
\code{\link{issueMessages}()},
\code{\link{issuePlot}()},
\code{\link{issuePrint}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{twoSeasonMessage}()},
\code{\link{zeroDateMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
