% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{timeTravelMessage}
\alias{timeTravelMessage}
\title{Return message if an issue date is after the file was submitted}
\usage{
timeTravelMessage(issue_assignments)
}
\arguments{
\item{issue_assignments}{An intermediate tibble in \code{\link{issueCheck}}}
}
\description{
The internal \code{timeTravelMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if \code{issue_date} values are after the date the file was submitted.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
