% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate data}
\usage{
validate(x)
}
\arguments{
\item{x}{The object created after tidying data with \code{\link{tidy}}}
}
\description{
After tidying the data with \code{\link{tidy}}, check to make sure the data don't have any erroneously repeated values.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
