% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate data}
\usage{
validate(x, type, all = FALSE)
}
\arguments{
\item{x}{The object created after cleaning data with \code{\link{clean}}}

\item{type}{Type of validation to perform. Acceptable values include:
\itemize{
\item vertical - Checks for repetition vertically in species and/or bag fields, grouped by dl_state and dl_date
\item horizontal - Checks for repetition horizontally, across each record
}}

\item{all}{Should all species groups be checked (TRUE)? If set to FALSE (default), then only ducks will be vertically checked and only ducks, geese, and coots_snipe will be horizontally checked.}
}
\description{
After cleaning the data with \code{\link{clean}}, check to make sure the data don't have any erroneously repeated values.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\seealso{
\code{\link[=investigate]{investigate()}} to look into vertical repetition.
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
