% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_hip.R
\name{write_hip}
\alias{write_hip}
\title{Modify corrected data table format and write as csv}
\usage{
write_hip(x, path)
}
\arguments{
\item{x}{The object created after correcting data with \code{\link{correct}}}

\item{path}{The file path and file name to write the final table}
}
\description{
After correcting errors in the data with \code{\link{correct}}, this final step will shape up the dataframe into a format ready for the database, and write the data to csv.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
