% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.R
\name{writeReport}
\alias{writeReport}
\title{Write an R markdown report from template}
\usage{
writeReport(path, type, yr, dl = NA, dir, file)
}
\arguments{
\item{path}{File path to the folder containing HIP .txt files}

\item{type}{Type of report. One of the following options may be supplied:
\itemize{
\item dl_report - for a download cycle summary report
\item season_report - for a summary of the entire HIP season}}

\item{yr}{The year in which the Harvest Information Program data were collected}

\item{dl}{Download cycle (when running a download report)}

\item{dir}{Folder in which to save the completed report}

\item{file}{What the report file should be named}
}
\description{
Create documentation with figures and tables that summarizes HIP download data at a download cycle or season total scale.
}
\references{
\url{https://github.com/USFWS/migbirdHarvestData}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
