% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.R
\name{writeReport}
\alias{writeReport}
\title{Write an R markdown report from template}
\usage{
writeReport(
  path,
  yr,
  dl = NA,
  corrected_path,
  future_path,
  past_path,
  dir,
  file
)
}
\arguments{
\item{path}{File path to the folder containing HIP .txt files}

\item{yr}{The year in which the Harvest Information Program data were collected}

\item{dl}{Download cycle (when running a download report)}

\item{corrected_path}{Path to directory that contains output .csv files}

\item{future_path}{Path to directory that contains future data .csv files}

\item{past_path}{Path to directory that contains past data .csv files}

\item{dir}{Folder in which to save the completed report}

\item{file}{What the report file should be named}
}
\description{
Create documentation with figures and tables that summarizes HIP download data at a download cycle scale.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
