% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeReport.R
\name{writeReport}
\alias{writeReport}
\title{Write an R markdown report from template}
\usage{
writeReport(raw_path, temp_path, future_path, past_path, year, dl, dir, file)
}
\arguments{
\item{raw_path}{File path to the folder containing HIP .txt files}

\item{temp_path}{Path to directory that contains output .csv files}

\item{future_path}{Path to directory that contains future data .csv files}

\item{past_path}{Path to directory that contains past data .csv files}

\item{year}{The year in which the Harvest Information Program data were collected}

\item{dl}{Download cycle}

\item{dir}{Folder in which to save the rendered report}

\item{file}{What the report file should be named (do not include an extension or suffix indicating file type, this is done automatically)}
}
\description{
Create documentation with figures and tables that summarizes HIP data at a download cycle scale.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
