% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/issuance.R
\name{zeroDateMessage}
\alias{zeroDateMessage}
\title{Return message for 00/00/0000 issue_date value(s)}
\usage{
zeroDateMessage(clean_data)
}
\arguments{
\item{clean_data}{The product of \code{\link{clean}}}
}
\description{
The internal \code{zeroDateMessage} function is used inside of
\code{\link{issueMessages}} and \code{\link{issueCheck}} to return a message
if "00/00/0000" values are detected in the \code{issue_date} field.
}
\seealso{
Other issuance functions: 
\code{\link{badDateMessage}()},
\code{\link{futureDateMessage}()},
\code{\link{invalidDateMessage}()},
\code{\link{issueAssign}()},
\code{\link{issueCheck}()},
\code{\link{issueDecide}()},
\code{\link{issueMessages}()},
\code{\link{issuePlot}()},
\code{\link{issuePlotDateLabel}()},
\code{\link{issuePrint}()},
\code{\link{pastDateMessage}()},
\code{\link{regYearEditMessage}()},
\code{\link{timeTravelMessage}()},
\code{\link{twoSeasonMessage}()}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
\concept{issuance functions}
