% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{zipCheck}
\alias{zipCheck}
\title{Check if zip codes are associated with the correct state}
\usage{
zipCheck(raw_df)
}
\arguments{
\item{raw_df}{The object created after reading in data with \code{\link{read_hip}}}
}
\description{
The internal \code{zipCheck} function checks to see if zip codes in hunter addresses match the address state.
}
\references{
\url{https://github.com/USFWS/migbirdHIP}
}
\author{
Abby Walter, \email{abby_walter@fws.gov}
}
